% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background-correct.r
\name{background.correct}
\alias{background.correct}
\title{Background correction}
\usage{
background.correct(rg, probes, offset = 15, verbose = F)
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{read.rg}()}.}

\item{offset}{Number to add to background corrected signal (Default: 15).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
Background corrected Cy5/Cy3 signal by \code{\link[limma]{normexp.signal}}.
}
\description{
Background correct Cy5/Cy3 signal of a Infinium HumanMethylation450 BeadChip.
}

