% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse35069-references.r
\name{create.gse35069.references}
\alias{create.gse35069.references}
\title{Defines cell type references "blood gse35069" and "blood gse35069 complete"
for estimating blood cell counts.
Both use methylation profiles from Reinius et al. 2012 (PMID 25424692)
for purified blood cell types.
The first is based on
six cell types: CD4T, CD8T, Mono, Bcell, NK, Gran.
The second is based on 
the same cell types but with Gran replaced by Neu and Eos.}
\usage{
create.gse35069.references(number.pcs = 5, temp.dir = NULL, verbose = F)
}
\description{
Defines cell type references "blood gse35069" and "blood gse35069 complete"
for estimating blood cell counts.
Both use methylation profiles from Reinius et al. 2012 (PMID 25424692)
for purified blood cell types.
The first is based on
six cell types: CD4T, CD8T, Mono, Bcell, NK, Gran.
The second is based on 
the same cell types but with Gran replaced by Neu and Eos.
}

