% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye-bias-correct.r
\name{dye.bias.correct}
\alias{dye.bias.correct}
\title{Dye bias correction}
\usage{
dye.bias.correct(rg, probes, intensity = 5000, verbose = F)
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{read.rg}()}.}

\item{probes}{Output from \code{\link{meffil.probe.info}()} compatible with \code{rg}.}

\item{intensity}{Intensity of both color channels after correct (Default: 5000).}
}
\value{
Cy5/Cy3 signals with average intensity equal to \code{intensity}.
}
\description{
Adjusts dye bias by scaling each color channel to have the same mean intensity.
}
\details{
Infinium HumanMethylation450 BeadChip
}

