% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{ewas.lm}
\alias{ewas.lm}
\title{ewas.lm}
\usage{
ewas.lm(variable, beta, covariates = NULL, weights = NULL,
  winsorize.pct = 0.05, verbose = F, phenotype.data = NULL,
  family = NULL)
}
\description{
performs EWAS using r's lm()/glm() functions
}
\details{
Test associations between \code{variable} and each row of \code{beta}
while adjusting for \code{covariates}
}
