% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{ewas}
\alias{ewas}
\title{Test associations between \code{variable} and each row of \code{beta}
while adjusting for \code{covariates} (fixed effects) and \code{batch} (random effect).
If \code{cell.counts} is not \code{NULL}, then it is assumed that
the methylation data is derived from samples with two cell types.
\code{cell.counts} should then be a vector of numbers
between 0 and 1 of length equal to \code{variable} corresponding
to the proportions of cell of a selected cell type in each sample.
The regression model is then modified in order to identify
associations specifically in the selected cell type (PMID: 24000956).}
\usage{
ewas(variable, beta, covariates = NULL, batch = NULL, weights = NULL,
  cell.counts = NULL, winsorize.pct = 0.05, robust = TRUE, rlm = FALSE,
  verbose = F)
}
\description{
Test associations between \code{variable} and each row of \code{beta}
while adjusting for \code{covariates} (fixed effects) and \code{batch} (random effect).
If \code{cell.counts} is not \code{NULL}, then it is assumed that
the methylation data is derived from samples with two cell types.
\code{cell.counts} should then be a vector of numbers
between 0 and 1 of length equal to \code{variable} corresponding
to the proportions of cell of a selected cell type in each sample.
The regression model is then modified in order to identify
associations specifically in the selected cell type (PMID: 24000956).
}
