% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-from-samplesheet.r
\name{extract.from.samplesheet}
\alias{extract.from.samplesheet}
\title{select columns identified by \code{names}
from the samplesheet used to create the \code{qc.objects}.}
\usage{
extract.from.samplesheet(qc.objects, names)
}
\description{
select columns identified by \code{names}
from the samplesheet used to create the \code{qc.objects}.
}

