% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.r
\name{filter.features}
\alias{filter.features}
\title{Get a subset of features from a predefined feature set.}
\usage{
filter.features(featureset, expr)
}
\arguments{
\item{featureset}{A name returned by \code{\link{meffil.list.featuresets()}}.}

\item{expr}{An expression using column names in the feature set data frame
to select specific features.}
}
\value{
A data frame listing all features satisfying \code{expr}.
}
\description{
Get a subset of features from a predefined feature set.
}
\examples{
## Obtain a list of features measuring methylation using Type II probes.
filter.features("450k", target=="methylation" & type == "ii")

}

