% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaphunter.r
\name{gaphunter}
\alias{gaphunter}
\title{gaphunter}
\usage{
gaphunter(object, threshold = 0.05, keepOutliers = FALSE,
  outCutoff = 0.01, verbose = TRUE)
}
\arguments{
\item{object}{a matrix or object of class: "GenomicRatioSet", "GenomicMethylSet", "MethylSet", or"RatioSet"}

\item{threshold}{(default = 0.05)}

\item{keepOutliers}{(default = FALSE)}

\item{outCutoff}{(default = 0.01)}

\item{verbose}{(default = TRUE)
gaphunter() function from:
https://epigeneticsandchromatin.biomedcentral.com/articles/10.1186/s13072-016-0107-z
copied with minor modification from https://raw.githubusercontent.com/kasperdanielhansen/minfi/master/R/gaphunter.R (2018-03-23)}
}
