% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{guess.batch.vars}
\alias{guess.batch.vars}
\title{Guess which columns in sample sheet are batch variables}
\usage{
guess.batch.vars(qc.quantiles)
}
\arguments{
\item{qc.quantiles}{Output from \link{meffil.normalize.quantiles}}
}
\value{
Array of variable names
}
\description{
Guess which columns in sample sheet are batch variables
}
\examples{
\dontrun{

}
}

