% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features.r
\name{guess.featureset}
\alias{guess.featureset}
\title{select a compatible featureset for the set of features.}
\usage{
guess.featureset(features)
}
\description{
select a compatible featureset for the set of features.
}

