% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-matrix.r
\name{impute.matrix}
\alias{impute.matrix}
\title{to impute missing values with row means
x <- impute.matrix(x,1)
to impute missing values with column means
x <- impute.matrix(x,2)}
\usage{
impute.matrix(x, margin = 1, fun = function(x) mean(x, na.rm = T))
}
\description{
to impute missing values with row means
x <- impute.matrix(x,1)
to impute missing values with column means
x <- impute.matrix(x,2)
}

