% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-references.r
\name{meffil.add.cell.type.reference}
\alias{meffil.add.cell.type.reference}
\title{Create a cell type reference object}
\usage{
meffil.add.cell.type.reference(name, M, U, cell.types, chip = NA,
  featureset = chip, number.sites = 50, specific.sites = NULL,
  number.quantiles = 500, subsets = NULL, object = NULL,
  verbose = F)
}
\arguments{
\item{name}{Character string providing the name of the reference.}

\item{M}{Matrix of methylated probe intensities (rows=CpG sites, columns=samples).}

\item{U}{Matrix of unmethylatd probe intensities (rows=CpG sites, columns=samples).}

\item{cell.types}{Vector of cell type names corresponding to sample \code{basename}s.}

\item{chip}{Name returned by \code{\link{meffil.list.chips()}} (Default: NA).}

\item{featureset}{Name returned by \code{\link{meffil.list.featuresets()}} (Default: chip).}

\item{number.sites}{Number of probes to characterise cell type methylation (Default: 50).
For each cell type, this number of probes with greater methylation than other cell types
and the same number with lesser methylation than the other cell types will be included.}

\item{specific.sites}{If not null (default), then \code{number.sites} is ignored and
the supplied site identifiers are used to differentiate between cell types instead
of those maximally different between the cell types within the reference.}

\item{number.quantiles}{Length of numeric sequence to specify probe intensity distributions
(Default: 500).}

\item{object}{Cell type reference previously created by this function.
If not \code{NULL}, then this reference is added
and all other function arguments are ignored (Default: NULL).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}
}
\value{
A list specifying a cell type reference object that can be used by
\code{\link{meffil.estimate.cell.counts}()} to estimate cell counts
in another dataset.
The object is a list containing:
- \code{beta} The normalized methylation values of sites
differentially methylated between cell types.
- \code{quantiles} The average quantiles of methylated and unmethylated signals
of probe sets defined by \code{subsets} (see below).
e.g. \code{quantiles[[name]]$M} provides the quantiles (\code{number.quantiles} quantiles)
of the probes specified by \code{subsets[[name]]}.
- \code{subsets} Probes on the microarray partitioned by relationship to CpG islands,
either in an island, in a shore or far from an island.
}
\description{
Create a cell type reference object for estimating cell counts
with the Infinium HumanMethylation450 BeadChip.
}
