% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.add.chip}
\alias{meffil.add.chip}
\title{Add a new chip for analysis.}
\usage{
meffil.add.chip(name, manifest, intersections = TRUE)
}
\arguments{
\item{name}{Name of the new chip.}

\item{manifest}{A data frame obtained by loading the Illumina
manifest into R.}
}
\value{
Assuming that \code{manifest} contains a satisfactory
set of columns, a new feature set and a new chip is made available.
Thus, \code{name} will be added to the vectors returned by
\code{\link{meffil.list.featuresets}()} and
\code{\link{meffil.list.chips}()}.

The manifest must contain the following columns:
\itemize{
\item{"IlmnID"}{character}
\item{"Name"}{character}
\item{"AddressA_ID"}{character}
\item{"AddressB_ID"}{character}
\item{"Infinium_Design_Type"}{values "I","II" or ""}
\item{"CHR"}{values "0"-"22", "M", "X" or "Y"}
\item{"MAPINFO"}{integer}
\item{"AlleleA_ProbeSeq"}{character}
\item{UCSC_RefGene_Name}{character}
\item{UCSC_RefGene_Accession}{character}
\item{UCSC_RefGene_Group}{character}
\item{"UCSC_CpG_Islands_Name"}{character}
\item{"Relation_to_UCSC_CpG_Island"}{character}
\item{"snp.exclude"}{logical}
}
}
\description{
Add a new chip for analysis.
}
