% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-references.r
\name{meffil.add.cnv.reference}
\alias{meffil.add.cnv.reference}
\title{Create a copy number reference object}
\usage{
meffil.add.cnv.reference(name, M, U, chip = NA, featureset = chip,
  object = NULL, verbose = T)
}
\arguments{
\item{name}{Character string providing the name of the reference.}

\item{M}{Matrix of methylated probe intensities (rows=CpG sites, columns=samples).}

\item{U}{Matrix of unmethylatd probe intensities (rows=CpG sites, columns=samples).}

\item{chip}{Name returned by \code{\link{meffil.list.chips()}} (Default: NA).}

\item{featureset}{Name returned by \code{\link{meffil.list.featuresets()}}
(Default: \code{chip}).}

\item{object}{A previously created copy number reference object created by this function.
If not \code{NULL}, then this reference is added with the given \code{name}
and all other function arguments are ignored (Default: NULL).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}
}
\value{
A list specifying a copy number reference object that can be used by
\code{\link{meffil.calculate.cnv}()} to estimate copy number variation in another dataset.
}
\description{
Create a copy number reference object for estimating copy number variation
with the Infinium HumanMethylation450 BeadChip.
}

