% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-references.r
\name{meffil.add.copynumber450k.references}
\alias{meffil.add.copynumber450k.references}
\title{Create copy number references from CopyNumber450kData}
\usage{
meffil.add.copynumber450k.references(verbose = T)
}
\description{
Two copy number references are created using data from the
Bioconductor CopyNumber450kData R package.
Reference "copynumber450k" is created using the "450k" feature set,
and reference "copynumber450k-common" is created using the "common" feature set
so it can be used with datasets with mixed 450K and EPIC chips.
}
