% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.all.features}
\alias{meffil.all.features}
\title{Feature information for all Illumina Beadchips}
\usage{
meffil.all.features()
}
\value{
A data frame listing all features.
}
\description{
Feature information for all Illumina Beadchips
}
