% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosomal-sites.r
\name{meffil.autosomal.subset}
\alias{meffil.autosomal.subset}
\title{Restrict to subset of features targetting autosomal CpG sites}
\usage{
meffil.autosomal.subset(features)
}
\arguments{
\item{features}{A vector of feature names.}
}
\value{
Subset of the input targetting autosomal CpG sites
}
\description{
Restrict to subset of features targetting autosomal CpG sites
}
