% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functional-normalization.r
\name{meffil.background.correct}
\alias{meffil.background.correct}
\title{Background correction}
\usage{
meffil.background.correct(rg, probes = meffil.probe.info(), offset = 15)
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{meffil.read.rg}()}.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{offset}{Number to add to background corrected signal (Default: 15).}
}
\value{
Background corrected Cy5/Cy3 signal by \code{\link[limma][normexp.signal}.
}
\description{
Background correct Cy5/Cy3 signal of a Infinium HumanMethylation450 BeadChip.
}

