% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basenames.r
\name{meffil.basenames}
\alias{meffil.basenames}
\title{IDAT file basenames}
\usage{
meffil.basenames(path, recursive = FALSE)
}
\arguments{
\item{path}{Directory containing the IDAT files.}

\item{recursive}{If \code{TRUE}, search for IDAT files in subdirectories as well
(Default: \code{FALSE}).}
}
\value{
Character vector of IDAT file basenames
(i.e. filenames with "_Grn.idat" and "_Red.idat" removed).
In other words, each identifies the Cy5 and Cy3 output files corresponding to a single microarray.
}
\description{
List IDAT file basenames in a given directory.
}

