% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.calculate.cnv}
\alias{meffil.calculate.cnv}
\title{Calculate CNVs from IDAT files}
\usage{
meffil.calculate.cnv(samplesheet, cnv.reference, chip = NA, verbose = FALSE,
  ...)
}
\arguments{
\item{samplesheet}{Output from \code{meffil.create.samplesheet}}

\item{cnv.reference}{Name returned by \code{\link{meffil.list.cnv.references}()}.}

\item{chip}{Name returned by \code{\link{meffil.list.chips()}} (Default: NA).}

\item{verbose}{Default = FALSE}

\item{...}{Extra parameters to be passed to \code{DNAcopy} for segmentation. See details.}
}
\value{
Dataframe of segmented results
}
\description{
Based on the algorithm developed in \code{R/CopyNumber450k} bioconductor package
}
\details{
The following default values are being used:
- trim = 0.1
- min.width = 5
- nperm = 10000
- alpha = 0.001
- undo.splits = "sdundo"
- undo.SD = 2
}

