% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-counts.r
\name{meffil.cell.count.estimates}
\alias{meffil.cell.count.estimates}
\title{Cell count estimates}
\usage{
meffil.cell.count.estimates(qc.objects)
}
\arguments{
\item{qc.objects}{List of objects obtained from \code{\link{meffil.qc}()}
or \code{\link{meffil.create.qc.object}()}.}
}
\description{
Cell count estimates
}
