% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-count-qc-plot.r
\name{meffil.cell.count.qc.plot}
\alias{meffil.cell.count.qc.plot}
\title{Cell count estimate quality plot}
\usage{
meffil.cell.count.qc.plot(count.objects, reference)
}
\arguments{
\item{count.objects}{A list of objects each obtained from
\code{\link{meffil.estimate.cell.counts}()}.}

\item{reference}{Object describing methylation profiles of purified cell populations
obtained from \code{\link{meffil.create.cell.type.reference}()}.}
}
\value{
Two \link{ggplot2} boxplot objects:
- \code{betas} Contains one box per sample or reference cell type
representing the distribution of methylation levels for the CpG sites used to
estimate cell counts.
- \code{counts} Contains one box per reference cell type
representing the distribution of cell count estimates across the samples.
}
\description{
Cell count estimate quality plot
}

