% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-count-qc-plots.r
\name{meffil.cell.count.qc.plots}
\alias{meffil.cell.count.qc.plots}
\title{Cell count estimate quality plot}
\usage{
meffil.cell.count.qc.plots(count.objects)
}
\arguments{
\item{count.objects}{A list of objects each obtained from
\code{\link{meffil.estimate.cell.counts}()}.}
}
\value{
Two \link{ggplot2} boxplot objects:
\itemize{
\item \code{betas} Contains one box per sample or reference cell type
representing the distribution of methylation levels for the CpG sites used to
estimate cell counts.
\item \code{counts} Contains one box per reference cell type
representing the distribution of cell count estimates across the samples.
}
}
\description{
Cell count estimate quality plot
}
