% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.controls}
\alias{meffil.cnv.controls}
\title{Create controls for calculating CNVs}
\usage{
meffil.cnv.controls(featureset = "450k", verbose = FALSE)
}
\arguments{
\item{featureset}{Name returned by \code{\link{meffil.list.featuresets()}} (Default: \code{"450k"}).
Must be compatible with the "450k" chip.}

\item{verbose}{(Default: FALSE)}
}
\value{
List with pre-calculated CNV data
}
\description{
Create controls for calculating CNVs
}

