% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.matrix}
\alias{meffil.cnv.matrix}
\title{Create matrix of CNV values}
\usage{
meffil.cnv.matrix(cnv, abs.thresh = 0.25, p.thresh = 0.05)
}
\arguments{
\item{cnv}{Output from \code{meffil.calculate.cnv}}

\item{abs.thresh}{Seg.mean threshold for CNV to be called}

\item{p.thresh}{Bonferroni adjusted p-value threshold for CNV to be called}
}
\value{
Matrix of ncpg x nsample
}
\description{
Create matrix of CNV values
}

