% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.r
\name{meffil.collapse.dups}
\alias{meffil.collapse.dups}
\title{Collapse duplicate probes}
\usage{
meffil.collapse.dups(beta, dup.fun = function(x) median(x, na.rm = T))
}
\arguments{
\item{beta}{Methylation matrix returned by
\code{\link{meffil.normalize.samples}()}.}

\item{dup.fun}{Function to collapse duplicate probes
(Default: median).}
}
\value{
The input matrix with duplicated probes
(i.e. row names identical after stripping everything
after the "_" character) replaced by summaries defined by
\code{dup.fun}.
}
\description{
Collapse duplicated probes by replacing them with a summary.
}
