% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute-normalization-object.r
\name{meffil.compute.normalization.object}
\alias{meffil.compute.normalization.object}
\title{Normalization object}
\usage{
meffil.compute.normalization.object(basename, number.quantiles = 500,
  dye.intensity = 5000, probes = meffil.probe.info(), verbose = F,
  detection.threshold = 0.01, bead.threshold = 3)
}
\arguments{
\item{basename}{IDAT file basename (see \code{\link{meffil.basenames}}).}

\item{number.quantiles}{Number of quantiles to compute for probe subset (Default: 500).}

\item{dye.intensity}{Reference intensity for scaling each color channel (Default: 5000).}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}

\item{detection.threshold}{Default value = 0.01.
All probes above this detection threshold detected.}

\item{bead.threshold}{Default value = 3.
All probes with less than this number of beads detected.}
}
\value{
List containing control probe information, probe summaries
and quantiles.
}
\description{
Create a normalization object for a given Infinium HumanMethylation450 BeadChip.
}

