% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functional-normalization.r
\name{meffil.control.matrix}
\alias{meffil.control.matrix}
\title{Infinium HumanMethylation450 BeadChip control matrix}
\usage{
meffil.control.matrix(objects, scale.matrix = T)
}
\arguments{
\item{objects}{A list of outputs from \code{\link{meffil.compute.normalization.object}()}.}

\item{scale.matrix}{(Default: TRUE).}
}
\value{
Matrix with one column per object consisting of control probe intensities.
If scale.matrix is `TRUE`, then intensities are converted to z-scores,
missing values are imputed (row mean) and values more than 3 standard deviations
truncated.
}
\description{
Matrix containing control probe intensities from the Infinium HumanMethylation450 BeadChip.
}

