% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-samplesheet.r
\name{meffil.create.samplesheet}
\alias{meffil.create.samplesheet}
\title{Create sample sheet if an Illumina one isn't available}
\usage{
meffil.create.samplesheet(path, basenames = meffil.basenames(path, recursive),
  recursive = FALSE, delim = "_")
}
\arguments{
\item{basenames}{Output from \code{\link{meffil.basenames}}}

\item{delim}{Optional delim character to separate \code{Sample_Name} into multiple columns. Default: "_"}
}
\value{
Sample sheet data frame
}
\description{
If necessary generates two columns necessary for some functions: \code{Sample_Name} and \code{Sex}
}

