% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dye-bias-correct.r
\name{meffil.dye.bias.correct}
\alias{meffil.dye.bias.correct}
\title{Dye bias correction}
\usage{
meffil.dye.bias.correct(rg, intensity = 5000, probes = meffil.probe.info())
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{meffil.read.rg}()}.}

\item{intensity}{Intensity of both color channels after correct (Default: 5000).}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}
}
\value{
Cy5/Cy3 signals with average intensity equal to \code{intensity}.
}
\description{
Adjusts dye bias by scaling each color channel to have the same mean intensity.
}
\details{
Infinium HumanMethylation450 BeadChip
}

