% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/estimate-cell-counts.r
\name{meffil.estimate.cell.counts}
\alias{meffil.estimate.cell.counts}
\title{Estimate cell counts from a reference}
\usage{
meffil.estimate.cell.counts(object, mu = NULL, reference,
  probes = meffil.probe.info(), verbose = F)
}
\arguments{
\item{object}{An object created by \code{\link{meffil.compute.normalization.object}()}.}

\item{mu}{(Optional) Methylated and unmethylated intensities for a sample
created by \code{\link{meffil.rg.to.mu}}.  If this is not supplied,
then \code{object} is used to create such an object from the corresponding IDAT files.
The intensities will have been background corrected and dye bias corrected.}

\item{reference}{Object describing methylation profiles of purified cell populations
obtained from \code{\link{meffil.create.cell.type.reference}()}.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}
}
\value{
A list:
- \code{counts} Cell count estimates.
- \code{beta} Normalized methylation levels of sites used to differentiate
between reference cell types.

Results should be nearly identical to \code{\link[minfi]{estimateCellCounts}()}.
}
\description{
Estimate cell type ratios from methylation profiles of purified cell populations
(Infinium HumanMethylation450 BeadChip).
}

