% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedgraph.r
\name{meffil.ewas.bedgraph}
\alias{meffil.ewas.bedgraph}
\title{Save EWAS effect estimates to bedgraph file}
\usage{
meffil.ewas.bedgraph(
  ewas.object,
  filename,
  analysis,
  name,
  description,
  header
)
}
\arguments{
\item{ewas.object}{Object returned by \code{\link{meffil.ewas()}}.}

\item{filename}{Filename for output, typically with a 'bed' extension.}

\item{analysis}{The particular EWAS analysis from which to obtain summary statistics.
This should be one of \code{names(ewas.object$analyses)}.}

\item{name}{Text name to be included in the bedgraph header.}

\item{description}{Text description to be included in the bedgraph header.}

\item{header}{Bedgraph header. The default header uses the name and description provided.}
}
\description{
Saves EWAS effect estimates  to a bedgraph file
for viewing on a genome browser.
More file format details can be found here:
https://genome.ucsc.edu/goldenPath/help/bedgraph.html
}
