% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ewas-variable-summaries.r
\name{meffil.ewas.covariate.associations}
\alias{meffil.ewas.covariate.associations}
\title{Describe associations between EWAS covariates and the variable of interest.}
\usage{
meffil.ewas.covariate.associations(ewas.object)
}
\arguments{
\item{ewas.object}{Output of \code{\link{meffil.ewas}()}.}
}
\value{
A data frame with one or more rows for each covariate.

If both the variable of interest and covariate are continuous or ordinal,
then the covariate uses one row showing the name, mean
and standard deviation of the covariate following the significance
of the association between the covariate and the variable of interest.

If the covariate is categorical, then there is additionally one row for
each level showing the mean and standard deviation of the variable
of interest for samples at that covariate level.

If the variable of interest is categorical but the covariate is not,
then there is one row for each variable level showing the
mean and standard deviation of the covariate at the given level.

If both the variable of interest and covariate are categorical,
then mean is replaced with the number of samples at each pair of
variable/categorical levels and standard deviation with the percentage.
P-values indicate the significance of association using Fisher's exact test.
}
\description{
Describe associations between EWAS covariates and the variable of interest.
}

