% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ewas.r
\name{meffil.ewas.cpg.plot}
\alias{meffil.ewas.cpg.plot}
\title{use comet R package to plot region nearby}
\usage{
meffil.ewas.cpg.plot(ewas.object, cpg, title = cpg, beta = NULL)
}
\description{
use comet R package to plot region nearby
}

