% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{meffil.ewas.lm}
\alias{meffil.ewas.lm}
\title{Epigenome-wide association study (lm)}
\usage{
meffil.ewas.lm(beta, variable, covariates = NULL, weights = NULL,
  family = NULL, isva = T, sva = T, n.sv = NULL, isva0 = F,
  isva1 = F, winsorize.pct = 0.05, outlier.iqr.factor = NA,
  most.variable = min(nrow(beta), 50000), featureset = NA,
  random.seed = 20161123, phenotype.data = NULL, verbose = F)
}
\arguments{
\item{beta}{Methylation levels matrix, one row per CpG site, one column per sample.}

\item{variable}{Independent variable vector.}

\item{covariates}{Covariates data frame to include in regression model,
one row per sample, one column per covariate (Default: NULL).}

\item{weights}{numeric vector of weights to provide to the lm()/glm() function}

\item{family}{to perform generalised linear modeling with r's glm() function}

\item{isva}{Apply Independent Surrogate Variable Analysis (ISVA) to the
methylation levels and include the resulting variables as covariates in a
regression model (Default: TRUE).}

\item{sva}{Apply Surrogate Variable Analysis (SVA) to the
methylation levels and covariates and include
the resulting variables as covariates in a regression model (Default: TRUE).}

\item{n.sv}{Number of surrogate variables to calculate (Default: NULL).}

\item{winsorize.pct}{Apply all regression models to methylation levels
winsorized to the given level. Set to NA to avoid winsorizing (Default: 0.05).}

\item{outlier.iqr.factor}{For each CpG site, prior to fitting regression models,
set methylation levels less than
\code{Q1 - outlier.iqr.factor * IQR} or more than
\code{Q3 + outlier.iqr.factor * IQR} to NA. Here IQR is the inter-quartile
range of the methylation levels at the CpG site, i.e. Q3-Q1.
Set to NA to skip this step (Default: NA).}

\item{most.variable}{Apply (Independent) Surrogate Variable Analysis to the 
given most variable CpG sites (Default: 50000).}

\item{featureset}{Name from \code{\link{meffil.list.featuresets}()}  (Default: NA).}

\item{random.seed}{Value with which to seed the pseudo random number generator for reproducible results}

\item{phenotype.data}{data frame containing the variable and covariates to be used in the}

\item{verbose}{Set to TRUE if status updates to be printed (Default: FALSE).}
}
\description{
Test association with each CpG site.
}
