% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ewas.r
\name{meffil.ewas.manhattan.plot}
\alias{meffil.ewas.manhattan.plot}
\title{Manhattan plot for EWAS}
\usage{
meffil.ewas.manhattan.plot(ewas.object, sig.threshold = 1e-07,
  title = "Manhattan plot")
}
\arguments{
\item{ewas.object}{Return object from \code{\link{meffil.ewas()}}.}

\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{title}{Title for the plot (Default: "Manhattan plot").}
}
\value{
\code{\link{ggplot}} showing the Manhattan plot.
}
\description{
Manhattan plot for EWAS
}

