% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ewas-report.r
\name{meffil.ewas.parameters}
\alias{meffil.ewas.parameters}
\title{Specify parameters for QC}
\usage{
meffil.ewas.parameters(sig.threshold = 1e-07, max.plots = 10)
}
\arguments{
\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{max.plots}{Maximum number of plots to generate (Default: 10).}
}
\value{
List of parameter values
}
\description{
Specify parameters for QC
}

