% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plot-ewas.r
\name{meffil.ewas.qq.plot}
\alias{meffil.ewas.qq.plot}
\title{QQ plot for EWAS}
\usage{
meffil.ewas.qq.plot(ewas.object, sig.threshold = 1e-07, sig.color = "red",
  title = "QQ plot", xlab = bquote(-log[10]("expected p-values")),
  ylab = bquote(-log[10]("observed p-values")))
}
\arguments{
\item{ewas.object}{Return object from \code{\link{meffil.ewas()}}.}

\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{sig.color}{Color for points corresponding to significant tests (Default: "red").}

\item{title}{Title for the plot (Default: "QQ plot").}

\item{xlab}{Label for the x-axis (Default: -log_10(expected p-values)).}

\item{ylab}{Label for the y-axis (Default: -log_10(observed p-values)).}
}
\value{
\code{\link{ggplot}} showing the QQ plot.
}
\description{
QQ plot for EWAS
}

