% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ewas-report.r
\name{meffil.ewas.report}
\alias{meffil.ewas.report}
\title{Generate EWAS report.}
\usage{
meffil.ewas.report(ewas.summary, output.file = "ewas-report.html",
  author = "Analyst", study = "Illumina methylation data", ...)
}
\arguments{
\item{ewas.summary}{Output from \code{meffil.ewas.summary}.}

\item{output.file}{Default = "ewas-report.html".
If the file extension is not .htm, .html, .HTM or .HTML then
output will be in markdown format.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{study}{Default = "Illumina methylation data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\description{
Generate HTML file that summarises the EWAS.
}

