% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ewas-variable-summaries.r
\name{meffil.ewas.sample.characteristics}
\alias{meffil.ewas.sample.characteristics}
\title{Describe EWAS samples using the variable of interest and covariates.}
\usage{
meffil.ewas.sample.characteristics(ewas.object)
}
\arguments{
\item{ewas.object}{Output of \code{\link{meffil.ewas}()}.}
}
\value{
A data frame with one row for each continuous or ordinal variable
and one row for each level of each categorical variable.
In the first case, each row provides the name, mean value and standard deviation of each variable.
In the second case (categorical), each row provides the name of the variable level and
the number of cases and percentage of cases at that level.
}
\description{
Describe EWAS samples using the variable of interest and covariates.
}

