% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ewas-report.r
\name{meffil.ewas.summary}
\alias{meffil.ewas.summary}
\title{Summarize EWAS results.}
\usage{
meffil.ewas.summary(ewas.object, beta, selected.cpg.sites = character(0),
  parameters = meffil.ewas.parameters(), verbose = T)
}
\arguments{
\item{ewas.object}{From \code{\link{meffil.ewas}()}.}

\item{selected.cpg.sites}{Vector of CpG site names to plot (Default: character(0)).}

\item{parameters}{Default = meffil.ewas.parameters(). List of parameter values. See \code{\link{meffil.ewas.parameters}()}.}
}
\value{
List
}
\description{
Generates variable and covariate summary tables, QQ plots, Manhattan plots, a list of associations,
plots of the strongest associations and plots of selected CpG sites.
}

