% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.r
\name{meffil.expand.norm.object}
\alias{meffil.expand.norm.object}
\title{Restore information needed for sample normalization}
\usage{
meffil.expand.norm.object(norm.object, original)
}
\arguments{
\item{norm.object}{An output from \code{\link{meffil.normalize.quantiles}()}.}

\item{qc.object}{The output of \code{\link{meffil.create.qc.object}()}
that was used to create \code{norm.object}.}
}
\value{
A normalized "QC object" identical to \code{norm.object}
but containing the full complement of information found in \code{qc.object}.
}
\description{
Restore information to normalized QC objects derived from shrunk QC objects
}
\examples{
\dontrun{
random.effects = c("Slide")

qc.objects1 = meffil.qc(samplesheet1)
dat1.objects = lapply(qc.objects1, meffil.shrink.qc.object, keep.vars=random.effects)

qc.objects2 = meffil.qc(samplesheet2)
dat2.objects = lapply(qc.objects2, meffil.shrink.qc.object, keep.vars=random.effects)

qc.objects = c(dat1.objects, dat2.objects)
norm.objects = meffil.normalize.quantiles(qc.objects, number.pcs=5, random.effects=random.effects)
dat1.norms = norm.objects[names(dat1.objects)]
dat2.norms = norm.objects[names(dat2.objects)]

norm.objects1 = mapply(meffil.expand.norm.object, dat1.norms, qc.objects1, SIMPLIFY=F)
meth1 = meffil.normalize.samples(norm.objects1)

norm.objects2 = mapply(meffil.expand.norm.object, dat2.norms, qc.objects2, SIMPLIFY=F)
meth2 = meffil.normalize.samples(norm.objects2)
}

}
