% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-genotypes.r
\name{meffil.extract.genotypes}
\alias{meffil.extract.genotypes}
\title{Extract genotype data from PLINK .raw files for Illumina 450K SNPs}
\usage{
meffil.extract.genotypes(filenames, verbose = F)
}
\arguments{
\item{filenames}{A vector of filenames of PLINK .raw files from which to extract
genotype data.}
}
\value{
Matrix with rows corresponding to SNPs, columns to samples and values
equal to 0, 1 or 2 corresponding to genotypes.
}
\description{
Extract genotype data from PLINK .raw files for Illumina 450K SNPs
}
\examples{
R> writeLines(meffil.snp.names("450k"), con="snp-names.txt")
shell> plink --bfile dataset --extract snp-names.txt --recodeA --out genotypes.raw --noweb
R> filenames <- "genotypes.raw"
R> genotypes <- meffil.extract.genotypes(filenames)

}

