% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.r
\name{meffil.gds.apply}
\alias{meffil.gds.apply}
\title{Return a vector or list of values obtained by applying a function
to the margins of a methylation or detection p-value matrix
stored in a GDS file.}
\usage{
meffil.gds.apply(
  gds.filename,
  bysite = T,
  type = c("list", "none", "integer", "double", "character", "logical", "raw"),
  FUN,
  sites = NULL,
  samples = NULL,
  ...
)
}
\arguments{
\item{gds.filename}{Name of GDS file generated by \code{\link{meffil.normalize.samples}()}}

\item{bysite}{If `TRUE`, then apply function to each CpG site (row),
otherwise to each sample (column) (Default: TRUE).}

\item{type}{returned value.}

\item{FUN}{the function to be applied.}

\item{sites}{Names of CpG sites to apply to, `NULL` means all sites (Default: NULL).}

\item{samples}{Names of samples to apply to, `NULL` means all samples (Default: NULL).}

\item{...}{}
}
\value{

}
\description{
Return a vector or list of values obtained by applying a function
to the margins of a methylation or detection p-value matrix
stored in a GDS file.
}
