% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-detection-pvalues.r
\name{meffil.gds.detection.pvalues}
\alias{meffil.gds.detection.pvalues}
\title{Retrieve detection p-values from GDS file}
\usage{
meffil.gds.detection.pvalues(gds.filename, sites = NULL, samples = NULL)
}
\arguments{
\item{gds.filename}{Name of GDS file generated by \code{\link{meffil.save.detection.pvalues}()}.}

\item{sites}{Names of CpG sites to load, if `NULL` then load all (Default: NULL).}

\item{samples}{Names of samples to load, if `NULL` then load all (Default: NULL).}
}
\value{
Matrix of methylation levels with rows corresponding to CpG sites
and columns to samples.  Rows restricted \code{sites} if not \code{NULL},
and columns restricted to \code{samples} if not \code{NULL}.
}
\description{
Retrieve detection p-values from GDS file
}
