% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.r
\name{meffil.gds.dims}
\alias{meffil.gds.dims}
\title{Retrieve methylation or detection p-value matrix row and column names}
\usage{
meffil.gds.dims(gds.filename)
}
\arguments{
\item{gds.filename}{Name of GDS file generated by \code{\link{meffil.normalize.samples}()} or \code{\link{meffil.save.detection.pvalues}()}.}
}
\value{
A list of two vectors,
the first providing the row names (CpG sites)
and the second providing the column names (sample identifiers).
}
\description{
Retrieve methylation or detection p-value matrix row and column names
}
