% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.r
\name{meffil.gds.methylation}
\alias{meffil.gds.methylation}
\title{Retrieve methylation levels from GDS file}
\usage{
meffil.gds.methylation(gds.filename, sites = NULL, samples = NULL)
}
\arguments{
\item{gds.filename}{Name of GDS file generated by \code{\link{meffil.normalize.samples}()}.}

\item{sites}{Names of CpG sites to load, if \code{NULL} then load all (Default: NULL).}

\item{samples}{Names of samples to load, if \code{NULL} then load all (Default: NULL).}
}
\value{
Matrix of methylation levels with rows corresponding to CpG sites
and columns to samples.  Rows restricted \code{sites} if not \code{NULL},
and columns restricted to \code{samples} if not \code{NULL}.
}
\description{
Retrieve methylation levels from GDS file
}
