% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-beta.r
\name{meffil.get.beta}
\alias{meffil.get.beta}
\title{Infinium HumanMethylation450 BeadChip methylation levels}
\usage{
meffil.get.beta(M, U, pseudo = 100)
}
\arguments{
\item{M}{Methylated signal matrix.}

\item{U}{Unmethylated signal matrix.}

\item{pseudo}{Value to add to the denominator to make the methylation estimate more stable.}
}
\value{
Matrix of 0..1 methylation level estimates.
Equal to methylated/(methylated + unmethylated + pseudo).
}
\description{
Compute beta values (methylation levels) from methylated/unmethylated signals
}

