% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-references.r
\name{meffil.get.cell.type.references}
\alias{meffil.get.cell.type.references}
\title{List of available cell type references}
\usage{
meffil.get.cell.type.references()
}
\description{
List of available cell type references
}

