% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cell-type-references.r
\name{meffil.get.current.cell.type.reference}
\alias{meffil.get.current.cell.type.reference}
\title{Currently selected cell type reference}
\usage{
meffil.get.current.cell.type.reference()
}
\description{
Currently selected cell type reference
}

