% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/features.r
\name{meffil.get.features}
\alias{meffil.get.features}
\title{Get a list of microarray features from a predefined feature set.}
\usage{
meffil.get.features(featureset)
}
\arguments{
\item{featureset}{A name returned by \code{\link{meffil.list.featuresets()}}.}
}
\value{
A data frame listing all features in the feature set.
}
\description{
Get a list of microarray features from a predefined feature set.
}

