% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/site-subsets.r
\name{meffil.get.sites}
\alias{meffil.get.sites}
\title{Get names of all CpG sites in the feature set.}
\usage{
meffil.get.sites(featureset = "450k")
}
\description{
Get names of all CpG sites in the feature set.
}

