% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-subsets.r
\name{meffil.get.typeii.sites}
\alias{meffil.get.typeii.sites}
\title{Get names of CpG sites corresponding to Infinium Type II probes
in the feature set.}
\usage{
meffil.get.typeii.sites(featureset = "450k")
}
\description{
Get names of CpG sites corresponding to Infinium Type II probes
in the feature set.
}
