% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gse35069-references.r
\name{meffil.gse35069.references}
\alias{meffil.gse35069.references}
\title{Purified blood cell type methylation references}
\usage{
meffil.gse35069.references()
}
\value{
A list of two objects,
each created by \code{\link{meffil.create.cell.type.reference}()}.
The first, named \code{standard}, is a reference based on
six cell types: CD4T, CD8T, Mono, Bcell, NK, Gran.
The second, named \code{complete}, is a reference based on
the same cell types but with Gran replaced by Neu and Eos.
Each of these objects can be passed as the \code{reference}
parameter to the function \code{\link{meffil.estimate.cell.counts}()}.
}
\description{
Methylation data from Reinius et al. 2012 (PMID 25424692)
estimating cell blood cell counts in methylation profiles.
}

