% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.list.chips}
\alias{meffil.list.chips}
\title{List of microarrays formats available.}
\usage{
meffil.list.chips()
}
\description{
By default, there is '450k' and 'epic'.
Additions can be made using \code{\link{meffil.add.chip}()}.
}

