% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-references.r
\name{meffil.list.cnv.references}
\alias{meffil.list.cnv.references}
\title{List of available copy number references}
\usage{
meffil.list.cnv.references()
}
\description{
List of available copy number references
}
