% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.list.featuresets}
\alias{meffil.list.featuresets}
\title{List of feature sets available.}
\usage{
meffil.list.featuresets()
}
\description{
Sets of features for individual platforms (e.g. "450k" for the Illumina HumanMethylation450 Beadchip)
as well as for mixed platforms (e.g. "450k:epic:epic2" for
combinations of Illumina HumanMethylation450, MethylationEPIC and MethylationEPICv2).
}
\details{
In most cases, a feature corresponds to the two probes
from which it's value is derived.  Each CpG represented
on the chip for example
corresponds to a single feature derived from a probe measuring
methylated signal and a second probe measuring unmethylated signal.

Each control feature corresponds to a unique control probe.
}
