% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-controls.r
\name{meffil.load.controls}
\alias{meffil.load.controls}
\title{Load control probes}
\usage{
meffil.load.controls(samplesheet, chip = NA, featureset = chip,
  verbose = F, ...)
}
\arguments{
\item{samplesheet}{Sample info (see \code{\link{meffil.read.samplesheet}} or \code{\link{meffil.create.samplesheet}}).}

\item{chip}{Name returned by \code{\link{meffil.list.chips()}} (Default: \code{NA}).}

\item{featureset}{Name returned by \code{\link{meffil.list.featuresets()}} (Default: \code{chip}).}

\item{verbose}{(Default: FALSE).}

\item{...}{Arguments to mclapply.}
}
\value{
List containing two elements: probes and values.
The probes item is a data frame describing the control probes.
The values item is a matrix providing the intensities of the
control probes for each samples (rows=probes, columns=samples).
}
\description{
Load control probes
}
