% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/load-detection-pvals.r
\name{meffil.load.detection.pvalues}
\alias{meffil.load.detection.pvalues}
\title{Load detection p-value matrix}
\usage{
meffil.load.detection.pvalues(qc.objects, max.bytes = 2^30 - 1, verbose = F,
  ...)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}.}
}
\value{
Matrix of probe detection p-values.
}
\description{
Load detection p-value matrix
}

