% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-raw-betas.r
\name{meffil.load.raw.data}
\alias{meffil.load.raw.data}
\title{Load raw beta matrix}
\usage{
meffil.load.raw.data(
  qc.objects,
  pseudo = 100,
  just.beta = T,
  max.bytes = 2^30 - 1,
  verbose = F,
  ...
)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{pseudo}{Value to add to the denominator to make the methylation
estimate more stable when calculating methylation levels (Default: 100).}

\item{just.beta}{If \code{TRUE}, then return just the methylation levels; otherwise,
return the methylated and unmethylated matrices (Default: TRUE).}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}.}
}
\value{
If \code{just.beta == TRUE}, the matrix of 
methylation levels between between 0 and 1
equal to methylated signal/(methylated + unmethylated signal + pseudo).
Otherwise, a list containing two matrices, the methylated and unmethylated signals.
}
\description{
Load raw beta matrix
}
