% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.from.betas.r
\name{meffil.normalization.parameters.from.betas}
\alias{meffil.normalization.parameters.from.betas}
\title{Specify parameters for testing normalization}
\usage{
meffil.normalization.parameters.from.betas(
  batch.pcs = 1:10,
  batch.threshold = 1e-50,
  colours = NULL
)
}
\arguments{
\item{batch.pcs}{Default = 1:10. Number of PCs to test against batch variables}

\item{batch.threshold}{Default = 1e-50. Which pvalue threshold to show in table}

\item{colours}{Colours to use for scatterplots.}
}
\value{
List of parameters
}
\description{
Specify parameters for testing normalization
}
\examples{
\dontrun{

}
}
