% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{meffil.normalization.parameters}
\alias{meffil.normalization.parameters}
\title{Specify parameters for testing normalization}
\usage{
meffil.normalization.parameters(qc.quantiles,
  variables = guess.batch.vars(qc.quantiles), control.pcs = 1:10,
  probe.pcs = 1:10, probe.range = 1000)
}
\arguments{
\item{qc.quantiles}{Output from \link{meffil.normalize.quantiles}}

\item{variables}{Default = guess.batch.vars(qc.quantiles). Which variables in sample sheet to test}

\item{control.pcs}{Default = 1:10. Number of control PCs to test against batch variables}

\item{probe.pcs}{Default = 1:10. Number of probe PCs to test against batch variables}

\item{probe.range}{Default = 1:1000. Number of probes to use to calculate PCs for actual CpGs}
}
\value{
List of parameters
}
\description{
Specify parameters for testing normalization
}
\examples{
\dontrun{

}
}

