% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{meffil.normalization.report}
\alias{meffil.normalization.report}
\title{Generate report on normalization performance}
\usage{
meffil.normalization.report(normalization.summary,
  output.file = "meffil.normalization.report.html", author = "Analyst",
  studyname = "IlluminaHuman450 data", ...)
}
\arguments{
\item{normalization.summary}{Output from \code{meffil.normalization.summary}.}

\item{output.file}{Default = "meffil.normalization.report.html"
If specified then a html report will be generated summarising the normalization.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{studyname}{Default = "IlluminaHuman450 data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{rmarkdown::render}}}
}
\description{
Generate HTML file that summarises the normalization.
}
\examples{
\dontrun{

}
}

