% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.from.betas.r
\name{meffil.normalization.summary.from.betas}
\alias{meffil.normalization.summary.from.betas}
\title{Perform tests to check normalization performance}
\usage{
meffil.normalization.summary.from.betas(
  pcs,
  parameters = meffil.normalization.parameters.from.betas(),
  samplesheet = samplesheet,
  variables = variables,
  verbose = TRUE
)
}
\arguments{
\item{pcs}{Output from \code{\link{meffil.methylation.pcs}()}
applied to the normalized methylation matrix}

\item{parameters}{Default = meffil.normalization.parameters.from.betas(). Report parameters.}

\item{samplesheet}{Default = NULL. Data frame of variables to compare to
principal components (\code{pcs}).  Must contain \code{nrow(pcs) == nrow(samplesheet)} rows.
Columns that are not factors are ignored.}

\item{variables}{Which variables in sample sheet to test}

\item{verbose}{Default = TRUE}
}
\value{
List of tables and graphs.
}
\description{
Perform tests to check normalization performance
}
\examples{
\dontrun{

}
}
