% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{meffil.normalization.summary}
\alias{meffil.normalization.summary}
\title{Perform tests to check normalization performance}
\usage{
meffil.normalization.summary(normalized.beta, qc.quantiles,
  parameters = meffil.normalization.parameters(qc.quantiles),
  verbose = TRUE)
}
\arguments{
\item{normalized.beta}{Output from \code{meffil.normalize.samples}}

\item{qc.quantiles}{Output from \link{meffil.normalize.quantiles}}

\item{parameters}{Default = meffil.normalization.parameters(qc.quantiles) List of parameters.}

\item{verbose}{Default = TRUE}
}
\value{
List of tables and graphs
}
\description{
Creates scree plot of PCs of control probes, tests for association of control probe PCs with batch variables, tests for association of normalized probes with batch variables
}
\examples{
\dontrun{

}
}

