% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functional-normalization.r
\name{meffil.normalize.dataset}
\alias{meffil.normalize.dataset}
\title{Functional normalization}
\usage{
meffil.normalize.dataset(path, recursive = F, filenames, number.pcs = 2,
  sex = NULL, probes = meffil.probe.info())
}
\arguments{
\item{path}{Directory containing the idat files.  Ignored if \code{filenames} is defined.}

\item{recursive}{If \code{TRUE}, any idat file in \code{path} or a subdirectory is included
in the normalization; otherwise, only those in the immediate directory are included
(Default: \code{FALSE}).}

\item{filenames}{Optional character vector
listing the idat files to include in the normalization. Filenames may omit
the "_Grn.idat"/"_Red.idat" suffix.}

\item{number.pcs}{Number of control matrix principal components to adjust for (Default: 2).}

\item{sex}{Optional character vector assigning a sex label ("M" or "F") to each sample.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}
}
\value{
Normalized beta matrix (rows = CG sites, columns = samples, values = 0..1).
}
\description{
Apply functional normalization to a set of Infinium HumanMethylation450 BeadChip IDAT files.
}
\details{
Fortin JP, Labbe A, Lemire M, Zanke BW, Hudson TJ, Fertig EJ, Greenwood CM, Hansen KD.
Functional normalization of 450k methylation array data
improves replication in large cancer studies.
Genome Biol. 2014 Dec 3;15(12):503. doi: 10.1186/s13059-014-0503-2.
PMID: 25599564
}

