% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/normalize-objects.r
\name{meffil.normalize.objects}
\alias{meffil.normalize.objects}
\title{Normalize objects}
\usage{
meffil.normalize.objects(objects, number.pcs = 2, sex.cutoff = -2,
  sex = NULL, verbose = F)
}
\arguments{
\item{objects}{A list of outputs from \code{\link{meffil.compute.normalization.object}()}.}

\item{number.pcs}{Number of control matrix principal components to adjust for (Default: 2).}

\item{sex}{Optional character vector assigning a sex label ("M" or "F") to each sample.}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
Same list as input with additional elements added for each sample
including normalized quantiles needed for normalizing each sample.
}
\description{
Normalize microarray quantiles using controls extracted (Infinium HumanMethylation450 BeadChip).
}

