% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-sample.r
\name{meffil.normalize.sample}
\alias{meffil.normalize.sample}
\title{Normalize Infinium HumanMethylation450 BeadChip}
\usage{
meffil.normalize.sample(object, probes = meffil.probe.info(), verbose = F)
}
\arguments{
\item{object}{A list element from output of \code{\link{meffil.normalize.objects}()}.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
List containing normalized methylated and unmethylated signals.
}
\description{
Normalize sample methylation data using normalized quantiles.
}
\examples{
path <- ...
basenames <- meffil.basenames(path)
norm.objects <- lapply(basenames, function(basename) {
  meffil.compute.normalization.object(basename)
})
norm.objects <- meffil.normalize.objects(norm.objects, number.pcs=2)
mu1 <- mefill.normalize.sample(norm.objects[[1]])
beta1 <- mefill.get.beta(mu1)
}

