% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functional-normalization.r
\name{meffil.normalize.samples}
\alias{meffil.normalize.samples}
\title{Normalize Infinium HumanMethylation450 BeadChips}
\usage{
meffil.normalize.samples(objects, probes = meffil.probe.info())
}
\arguments{
\item{objects}{A list or sublist returned by \code{\link{meffil.normalize.objects}()}.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}
}
\value{
Normalized methylated and unmethylated signals for each object.
}
\description{
Normalize a set of samples using their normalization objects.
}

