% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{meffil.pcs}
\alias{meffil.pcs}
\title{Calculate control probe PCs}
\usage{
meffil.pcs(qc.objects, fixed.effects = NULL, random.effects = NULL)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}
}
\value{
PCA of control probes
}
\description{
Calculate control probe PCs
}

