% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.beadnum.cpgs}
\alias{meffil.plot.beadnum.cpgs}
\title{Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample}
\usage{
meffil.plot.beadnum.cpgs(samplesheet, norm.objects, threshold = 0.05)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}

\item{threshold}{Cut off value for proportion of samples with poor detection p values. Default 0.05.}
}
\value{
Data frame of results plus plot
}
\description{
Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample
}
\examples{
\dontrun{

}
}

