% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.beadnum.manhattan}
\alias{meffil.plot.beadnum.manhattan}
\title{Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample}
\usage{
meffil.plot.beadnum.manhattan(samplesheet, norm.objects)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}
}
\value{
Data frame of results plus plot
}
\description{
Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample
}
\examples{
\dontrun{

}
}

