% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.beadnum}
\alias{meffil.plot.beadnum}
\title{Plot number of beads per sample}
\usage{
meffil.plot.beadnum(samplesheet, norm.objects)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}
}
\value{
Data frame of results plus plot
}
\description{
Plot number of beads per sample
}
\examples{
\dontrun{

}
}

