% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.beadnum.samples}
\alias{meffil.plot.beadnum.samples}
\title{Plot number of beads per sample}
\usage{
meffil.plot.beadnum.samples(samplesheet, norm.objects, threshold = 0.05,
  colour.code = NULL)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}

\item{threshold}{Cut off value for proportion of CpGs with low bead numbers. Default 0.05}

\item{colour.code}{Array of length n samples to colour code points. Defaults to NULL}
}
\value{
Data frame of results plus plot
}
\description{
Plot number of beads per sample
}
\examples{
\dontrun{

}
}

