% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{meffil.plot.control.batch}
\alias{meffil.plot.control.batch}
\title{Test for association of control matrix probes with known batch variables}
\usage{
meffil.plot.control.batch(qc.quantiles, npcs = 1:10,
  variables = guess.batch.vars(qc.quantiles), verbose = TRUE)
}
\arguments{
\item{qc.quantiles}{From \code{meffil.normalize.objects}}

\item{pcs}{Which PCs to plot. Default first 10}

\item{variables.}{Default = guess.batch.vars(qc.quantiles). Array spacifying column names in samplesheet to test for association with control matrix PCs}

\item{verbose=T}{Print progress messages?}
}
\value{
Data frame of results plus plot
}
\description{
Performs association of each of \code{n} PCs calculated from most control matrix against each of \code{m} measured batch variables
}
\examples{
\dontrun{

}
}

