% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.plot.control.scree}
\alias{meffil.plot.control.scree}
\title{Plot scree plot of control matrix}
\usage{
meffil.plot.control.scree(norm.objects)
}
\arguments{
\item{norm.objects}{From \code{meffil.normalize.quantiles}}
}
\value{
Data frame of results plus plot
}
\description{
Plot scree plot of control matrix
}
\examples{
\dontrun{

}
}

