% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.controlmeans}
\alias{meffil.plot.controlmeans}
\title{Plot the means of control probes for each sample and for each control probe type}
\usage{
meffil.plot.controlmeans(qc.objects,
  control.categories = names(qc.objects[[1]]$controls), colour.code = NULL,
  outlier.sd = 5)
}
\arguments{
\item{qc.objects}{From \code{meffil.normalize.objects}}

\item{control.categories}{Which control probe categories to plot. Defaults to all available}

\item{colour.code}{Array of length n samples to colour code points. Defaults to NULL}

\item{outlier.sd}{Cut off for declaring outliers. Default = 5}
}
\value{
Data frame of results plus plot
}
\description{
Plot the means of control probes for each sample and for each control probe type
}
\examples{
\dontrun{

}
}

