% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.detectionp.cpgs}
\alias{meffil.plot.detectionp.cpgs}
\title{Manhattan plot of detection pval per probe - percentage with pvalue < 0.01}
\usage{
meffil.plot.detectionp.cpgs(qc.objects, threshold = 0.05)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}

\item{threshold}{Cut off value for proportion of samples with poor detection p values. Default 0.05.}
}
\value{
Data frame of results plus plot
}
\description{
Manhattan plot of detection pval per probe - percentage with pvalue < 0.01
}
\examples{
\dontrun{

}
}

