% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.detectionp.manhattan}
\alias{meffil.plot.detectionp.manhattan}
\title{Manhattan plot of detection pval per probe - percentage with pvalue < 0.01}
\usage{
meffil.plot.detectionp.manhattan(samplesheet, norm.objects)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}
}
\value{
Data frame of results plus plot
}
\description{
Manhattan plot of detection pval per probe - percentage with pvalue < 0.01
}
\examples{
\dontrun{

}
}

