% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.genotypes}
\alias{meffil.plot.genotypes}
\title{Plot SNP beta and sample genotype concordances}
\usage{
meffil.plot.genotypes(qc.objects, genotypes = NULL, sample.threshold = 0.9,
  snp.threshold = 0.99)
}
\arguments{
\item{qc.objects}{Output from \code{\link{meffil.qc}()}.}

\item{genotypes}{Optional output from \code{\link{meffil.extract.genotypes}()}.
Sample genotypes are matched to sample qc.objects using
\code{colnames(genotypes)} and \code{names(qc.objects)}.}

\item{sample.threshold}{Concordance threshold below which the Illumina 450K
and genetic profiles for a sample are deemed a mismatch (Default: 0.9).}

\item{snp.threshold}{Concordance threshold below which the Illumina 450K
and genetic profiles for a SNP are deemed a mismatch (Default: 0.99).}
}
\value{
A list consisting of:
- \code{graphs} A list of \link{ggplot2} objects.  The first \code{snp.betas}
plots the beta distributions of each SNP probe in the microarray.
The second and third plots are added only if the \code{genotypes} parameter
is not \code{NULL}. The second plot shows the distribution of SNP concordances,
and the third plot shows the distribution of sample concordances.
- \code{tabs} Contains two data frames if the
\code{genotypes} parameter is not \code{NULL}.  The first \code{samples}
lists the concordances of each sample, the second \code{snps} lists the
concordances of each SNP.
}
\description{
Plot SNP beta and sample genotype concordances
}

