% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.meth.unmeth}
\alias{meffil.plot.meth.unmeth}
\title{Plot average methylated vs unmethylated levels for each individuals}
\usage{
meffil.plot.meth.unmeth(qc.objects, outlier.sd = 3, colour.code = NULL)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}

\item{outlier.sd}{Cut off for declaring outliers. Default = 3}

\item{colour.code}{Array of length n samples to colour code points. Defaults to NULL}
}
\value{
Data frame of results plus plot
}
\description{
plot raw control probes and fit linear regression, remove samples that have sd(y - yhat) > mean*3
}
\examples{
\dontrun{

}
}

