% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/post-process-plots.r
\name{meffil.plot.pc.batch}
\alias{meffil.plot.pc.batch}
\title{Plot each PC against each of chiprow, chip column, plate, slide IF there is a significant lm}
\usage{
meffil.plot.pc.batch(samplesheet, norm.objects, threshold = 0.05,
  pcs = 1:10)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}

\item{threshold}{P-value threshold for whether to plot. Default 0.05}

\item{pcs}{Which PCs to plot. Default first 10}
}
\value{
Data frame of results plus plot
}
\description{
Plot each PC against each of chiprow, chip column, plate, slide IF there is a significant lm
}
\examples{
\dontrun{

}
}

