% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.from.betas.r
\name{meffil.plot.probe.batch.from.betas}
\alias{meffil.plot.probe.batch.from.betas}
\title{Test normalized betas for association with known batch variables}
\usage{
meffil.plot.probe.batch.from.betas(
  samplesheet,
  variables,
  pcs,
  batch.threshold = batch.threshold,
  cols = NULL,
  verbose = T
)
}
\arguments{
\item{variables}{Which variables in sample sheet to test}

\item{pcs}{Output from \code{\link{meffil.methylation.pcs}()}
applied to the normalized methylation matrix}

\item{samplesheet.}{Data frame containing variables to test for association
with PCs. Must have \code{nrow(pcs) == nrow(samplesheet)}.}

\item{verbose=T}{Print progress messages?}
}
\value{
List of table of results and graph
}
\description{
Performs association of each of \code{n} PCs calculated from most variable CpG sites (after normalization) against each of \code{m} measured batch variables
}
\examples{
\dontrun{

}
}
