% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.plot.probe.batch}
\alias{meffil.plot.probe.batch}
\title{Test normalized betas for association with known batch variables}
\usage{
meffil.plot.probe.batch(normalized.beta, norm.objects, npcs = 1:10,
  variables = guess.batch.vars(norm.objects), batch.threshold = 1e-50,
  probe.range = 5000, cols = NULL, verbose = T)
}
\arguments{
\item{normalized.beta}{Output from \code{meffil.normalize.samples}}

\item{norm.objects}{Output from \code{meffil.normalize.quantiles}}

\item{npcs}{Default = 1:10. Which CpG PCs to test}

\item{variables}{Default = guess.batch.vars(norm). Which variables in sample sheet to test}

\item{probe.range}{Default = 1:1000. How many probes to be used in calculating PCs}

\item{verbose=T}{Print progress messages?}
}
\value{
List of table of results and graph
}
\description{
Performs association of each of \code{n} PCs calculated from most variable CpG sites (after normalization) against each of \code{m} measured batch variables
}
\examples{
\dontrun{

}
}

