% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.sex}
\alias{meffil.plot.sex}
\title{Plot predicted sex}
\usage{
meffil.plot.sex(qc.objects, outlier.sd = 3)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}
}
\value{
Data frame of results plus plot
}
\description{
Plot predicted sex
}
\examples{
\dontrun{

}
}
