% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.pre.processing}
\alias{meffil.pre.processing}
\title{Generate lists of bad probes and bad samples}
\usage{
meffil.pre.processing(samplesheet, norm.objects, colour.code,
  control.categories = names(norm.objects[[1]]$controls),
  sex.outlier.sd = 3, meth.unmeth.outlier.sd = 3,
  control.means.outlier.sd = 5, detectionp.samples.threshold = 0.05,
  beadnum.samples.threshold = 0.05, detectionp.cpgs.threshold = 0.05,
  beadnum.cpgs.threshold = 0.05)
}
\arguments{
\item{samplesheet}{From \code{\link{read.450k.sheet}()}}

\item{norm.objects}{From \code{\link{meffil.normalize.objects}()}}

\item{colour.code}{See above}

\item{control.categories}{See above}

\item{sex.outlier.sd}{See above}

\item{meth.unmeth.outlier.sd}{See above}

\item{control.means.outlier.sd}{See above}

\item{detectionp.samples.threshold}{See above}

\item{beadnum.samples.threshold}{See above}

\item{detectionp.cpgs.threshold}{See above}

\item{beadnum.cpgs.threshold}{See above}
}
\value{
List
}
\description{
Generate lists of bad probes and bad samples
}
\examples{
\dontrun{

}
}

