% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.probe.info}
\alias{meffil.probe.info}
\title{Obtain a list of probes for a given feature set (chip).}
\usage{
meffil.probe.info(featureset, chip = featureset)
}
\arguments{
\item{featureset}{Name returned by \code{\link{meffil.list.featuresets()}}.}

\item{chip}{Name returned by \code{\link{meffil.list.chips()}} (Default: \code{featureset}).}
}
\value{
A data frame with one row per probe.  The full set of probes
for a chip is returned if \code{chip == featureset}; otherwise,
the probes are restricted to those corresponding to features in the feature set.
}
\description{
Obtain a list of probes for a given feature set (chip).
}

