% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.qc.summary}
\alias{meffil.qc.summary}
\title{Generate lists of bad probes and bad samples}
\usage{
meffil.qc.summary(qc.objects, parameters = meffil.qc.parameters())
}
\arguments{
\item{qc.objects}{From \code{meffil.normalize.objects}}

\item{parameters}{Default = meffil.qc.parameters(). List of parameter values. See \code{\link{meffil.qc.parameters}}}
}
\value{
List
}
\description{
Generate lists of bad probes and bad samples
}
\examples{
\dontrun{

}
}

