% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/quantile-normalize-sample.r
\name{meffil.quantile.normalize.sample}
\alias{meffil.quantile.normalize.sample}
\title{Quantile normalize sample}
\usage{
meffil.quantile.normalize.sample(object, mu = NULL, subsets, quantiles,
  dye.intensity = 5000, probes = meffil.probe.info(), verbose = F)
}
\arguments{
\item{object}{An object created by \code{\link{meffil.compute.normalization.object}()}.}

\item{mu}{(Optional) Methylated and unmethylated intensities for a sample
created by \code{\link{meffil.rg.to.mu}}.  If this is not supplied,
then \code{object} is used to create such an object from the corresponding IDAT files.
The intensities will have been background corrected and dye bias corrected.}

\item{subsets}{Named list of probe sets (see \code{quantiles}).}

\item{quantiles}{Named list of numeric sequences defining the target distributions
of probes in corresponding sets specified by \code{subsets}.
For a set of probes \code{subset[[name]]}, \code{quantiles[[name]]$M}
is an ordered sequence
of numbers defining the quantiles of the target distribution
the methylated signal of the probes.  Similarly
\code{quantiles[[name]]$U} defines the target distribution of their
unmethylated signal distribution.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
\code{mu} modified so that for probe set \code{subsets[[name]]},
\code{mu$M[subsets[[name]]]} has the distribution defined by \code{quantiles[[name]]$M}
and \code{mu$U[subsets[[name]]]} has the distribution defined by \code{quantiles[[name]]$U}.
}
\description{
Normalize quantiles of a sample to supplied quantiles (Infinium HumanMethylation450 BeadChip).
}

