% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.remove.ids}
\alias{meffil.remove.ids}
\title{Remove samples from QC objects}
\usage{
meffil.remove.ids(qc.objects, idlist.remove)
}
\arguments{
\item{qc.objects}{Output from \code{\link{meffil.qc}}}

\item{idlist.remove}{Array of Sample_Name IDs to be removed}
}
\value{
qc.objects with samples removed
}
\description{
Remove samples from QC objects
}
\examples{
\dontrun{

}
}

