% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.remove.samples}
\alias{meffil.remove.samples}
\title{Remove samples from QC objects}
\usage{
meffil.remove.samples(qc.objects, sample.ids)
}
\arguments{
\item{qc.objects}{Output from \code{\link{meffil.qc}}}

\item{sample.ids}{Array of sample.name IDs to be removed}
}
\value{
qc.objects with samples removed
}
\description{
Remove samples from QC objects
}
\examples{
\dontrun{

}
}

