% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rg-to-mu.r
\name{meffil.rg.to.mu}
\alias{meffil.rg.to.mu}
\title{Convert to methylated/unmethylated signal}
\usage{
meffil.rg.to.mu(rg, probes = meffil.probe.info(), verbose = F)
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{meffil.read.rg}()}.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then print time function started (Default: \code{FALSE}).}
}
\value{
List containing methylated and unmethylated signals.
}
\description{
Converts Cy5/Cy3 signals to methylated/unmethylated signals from a
Infinium HumanMethylation450 BeadChip.
}

