% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-detection-pvalues.r
\name{meffil.save.detection.pvalues}
\alias{meffil.save.detection.pvalues}
\title{Save detection p-value matrix to GDS file}
\usage{
meffil.save.detection.pvalues(
  qc.objects,
  gds.filename = NULL,
  max.bytes = 2^30 - 1,
  verbose = F,
  ...
)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{gds.filename}{If not \code{NULL} (default), then saves the output to a GDS (Genomic Data Structure).
This is for cases where the output is too large to fit into main memory.}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}.}
}
\value{
Matrix of probe detection p-values.
If \code{gds.filename} is not \code{NULL}, then
the output is saved to the GDS file rather than retained
in memory and returned.
The library 'gdsfmt' must be installed in this case.
}
\description{
Save detection p-value matrix to GDS file
}
