% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-type-references.r
\name{meffil.set.current.cell.type.reference}
\alias{meffil.set.current.cell.type.reference}
\title{Select cell type reference for cell count estimates}
\usage{
meffil.set.current.cell.type.reference(name)
}
\arguments{
\item{name}{Name of cell type reference previously created by
\code{\link{meffil.create.cell.type.reference}()}.

Predefined references include "blood gse35069" and "blood gse35069 complete".
Both use methylation profiles from Reinius et al. 2012 (PMID 25424692)
for purified blood cell types.
The first is based on
six cell types: CD4T, CD8T, Mono, Bcell, NK, Gran.
The second is based on
the same cell types but with Gran replaced by Neu and Eos.}
}
\description{
Select cell type reference for cell count estimates
}

