% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe-info.r
\name{meffil.set.probe.info}
\alias{meffil.set.probe.info}
\title{Change probe type and location annotation}
\usage{
meffil.set.probe.info(probe.info)
}
\description{
This function will fail if expected columns are missing.
See the existing annotation by calling \code{\link{meffil.probe.info}()}.
}

