% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.r
\name{meffil.shrink.qc.object}
\alias{meffil.shrink.qc.object}
\title{Shrink quality control object}
\usage{
meffil.shrink.qc.object(qc.object, keep.vars = NULL)
}
\arguments{
\item{qc.object}{An output from \code{\link{meffil.create.qc.object}()}.}

\item{keep.vars}{Samplesheet variables to retain if the object will be
used to normalize quantiles using \code{\link{meffil.normalize.quantiles}()}
(Default: NULL).}
}
\value{
A "QC object" identical to \code{qc.object} but containing
only information that is needed for normalizing quantiles.
}
\description{
Remove any information not absolutely needed for normalizing quantiles
}
\examples{
\dontrun{
random.effects = c("Slide")

qc.objects1 = meffil.qc(samplesheet1)
dat1.objects = lapply(qc.objects1, meffil.shrink.qc.object, keep.vars=random.effects)

qc.objects2 = meffil.qc(samplesheet2)
dat2.objects = lapply(qc.objects2, meffil.shrink.qc.object, keep.vars=random.effects)

qc.objects = c(dat1.objects, dat2.objects)
norm.objects = meffil.normalize.quantiles(qc.objects, number.pcs=5, random.effects=random.effects)
dat1.norms = norm.objects[names(dat1.objects)]
dat2.norms = norm.objects[names(dat2.objects)]

norm.objects1 = mapply(meffil.expand.norm.object, dat1.norms, qc.objects1, SIMPLIFY=F)
meth1 = meffil.normalize.samples(norm.objects1)

norm.objects2 = mapply(meffil.expand.norm.object, dat2.norms, qc.objects2, SIMPLIFY=F)
meth2 = meffil.normalize.samples(norm.objects2)
}

}
