% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-betas.r
\name{meffil.snp.names}
\alias{meffil.snp.names}
\title{Obtain the list of identifiers for the SNPs on the microarray.}
\usage{
meffil.snp.names(featureset = "450k")
}
\arguments{
\item{featureset}{Name from \code{\link{meffil.list.featuresets}()} (Default: "450k").}
}
\description{
Obtain the list of identifiers for the SNPs on the microarray.
}

