% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas-variable-summaries.r
\name{meffil.summarize.relationship}
\alias{meffil.summarize.relationship}
\title{Describe the relationship between two variables.}
\usage{
meffil.summarize.relationship(vars)
}
\arguments{
\item{vars}{A data frame with at least two columns.
The first two columns will be compared.}
}
\value{
A list whose elements depends on the types of the two variables.
In each case, the list contains the following elements:
\describe{
\item{var1}{Name of the first variable, i.e. colnames(vars)[1].}
\item{var2}{Name of the second variable.}
\item{r}{Spearman's correlation between the variables. This may be meaningless if one variable is an unordered factor.}
\item{r.p}{P-value corresponding to the correlation between the variables.}
\item{output}{The contents of the list formatted to be printed as markdown text.}
\item{plot}{A plot (ggplot2) visualizing the relationship.}
}

If both variables are factors, then the list will include a frequency table (\code{freq})
and a corresponding matrix of p-values (\code{p.values}) obtained using Fisher's test to test for
enrichment in each cell of the frequency table.

If one variable is a factor and the other numeric, then list will include
the F-statistic (\code{F.stat}) and p-value (\code{p.value}) from one-way analysis of variance.
There will also be a data frame (\code{cases}) with each row providing statistics from a t-test
comparing the numeric variable within and without each level of the factor variable.

If both variables are numeric, then the list will include
the F-statistic (\code{F.stat}) and p-value (\code{p.value})
from the linear model fitting the variables.
}
\description{
Describe the relationship between two variables.
}
