% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(beta, outlier.iqr.factor)
}
\arguments{
\item{beta}{Methylation levels matrix, one row per CpG site, one column per sample.}

\item{outlier.iqr.factor}{For each CpG site, prior to fitting regression models,
set methylation levels less than
\code{Q1 - outlier.iqr.factor * IQR} or more than
\code{Q3 + outlier.iqr.factor * IQR} to NA. Here IQR is the inter-quartile
range of the methylation levels at the CpG site, i.e. Q3-Q1.
Set to NA to skip this step (Default: NA)}
}
\description{
selects outliers based on the value of "outlier.iqr.factor"
}
