% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{pca.to.design.matrix}
\alias{pca.to.design.matrix}
\title{From PCA applied to the raw control matrix,
derive the corresponding design matrix using the first \code{number.pcs} principal components.}
\usage{
pca.to.design.matrix(pca.ret, number.pcs, fixed.effects = NULL,
  random.effects = NULL)
}
\description{
From PCA applied to the raw control matrix,
derive the corresponding design matrix using the first \code{number.pcs} principal components.
}
