% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{predict.design.matrix}
\alias{predict.design.matrix}
\title{Given a partition of the QC objects into two sets,
use the design matrix from one partition to predict the design
matrix for the other partition.}
\usage{
\method{predict}{design.matrix}(qc.objects, number.pcs, new.idx,
  fixed.effects = NULL, random.effects = NULL)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{number.pcs}{Number of principal components to include in the design matrix (Default: all).}

\item{new.idx}{Indices into \code{qc.objects} that identify the objects for which
the design matrix will be predicted.  The initial design matrix will be derived from
the other objects.}

\item{fixed.effects}{Names of columns in samplesheet that should be included as fixed effects
along with control matrix principal components (Default: NULL).}

\item{random.effects}{Names of columns in samplesheet that should be included as random effects
(Default: NULL).}
}
\value{
Design matrix with one column for the \code{new.idx} partition predicted from the
first \code{number.pcs} prinicipal components of the objects.
}
\description{
Given a partition of the QC objects into two sets,
use the design matrix from one partition to predict the design
matrix for the other partition.
}

