% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/read-rg.r
\name{read.rg}
\alias{read.rg}
\title{Read Infinium HumanMethylation450 BeadChip.}
\usage{
read.rg(basename, verbose = F)
}
\arguments{
\item{basename}{IDAT file basename (see \code{\link{meffil.basenames}}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
List containing raw Cy5 ('R') and Cy3 ('G') data including
the intensity mean, intensity standard deviation and number of contributing beads.
}
\description{
Reads Cy5 and Cy3 files for a given Infinium HumanMethylation450 BeadChip.
}

