% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rg-to-mu.r
\name{rg.to.mu}
\alias{rg.to.mu}
\title{Convert to methylated/unmethylated signal}
\usage{
rg.to.mu(rg)
}
\arguments{
\item{rg}{Cy5/Cy3 signal generated by \code{\link{read.rg}()}.}
}
\value{
List containing methylated and unmethylated signals.
}
\description{
Converts Cy5/Cy3 signals to methylated/unmethylated signals from a
Infinium HumanMethylation450 BeadChip.
}

