% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rmd2html.r
\name{rmd2html}
\alias{rmd2html}
\title{Knit report using working environment and convert to HTML}
\usage{
rmd2html(input.filename, output.filename, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}

\item{normalization.summary}{Output from \code{meffil.normalization.summary}.}

\item{output.file}{Default = "meffil.normalization.report.html"
If specified then a html report will be generated summarising the normalization.}
}
\description{
Warning: It is quite likely that this will be called within an RMD file
implying a recursive call to knit(). this will generate "duplicate label"
errors for unlabelled chunks. To avoid this, all RMD files in this
package should contain only named chunks.
Supposedly this error can also be avoided by setting the following option:
     options(knitr.duplicate.label='allow')
I tried this but it didn't seem to help.
}

