% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\name{save.globals}
\alias{save.globals}
\title{called by ../data-raw/globals.r to save generated global variables to Rdata files
for loading whenever the package is loaded.}
\usage{
save.globals(dir)
}
\description{
called by ../data-raw/globals.r to save generated global variables to Rdata files
for loading whenever the package is loaded.
}
