% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplify-variable.r
\name{simplify.variable}
\alias{simplify.variable}
\title{Convert a variable to a format that can be included in a design matrix.
Characters and unordered factors are converted to model matrices with binary dummy variables.
Ordered factors are converted to integers corresponding to ordered levels.
Other variables are left asis.}
\usage{
simplify.variable(v)
}
\description{
Convert a variable to a format that can be included in a design matrix.
Characters and unordered factors are converted to model matrices with binary dummy variables.
Ordered factors are converted to integers corresponding to ordered levels.
Other variables are left asis.
}

