% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{svaIsva}
\alias{svaIsva}
\title{svaIsva}
\usage{
svaIsva(beta.sva, covariates, random.seed, mod, n.sv, mode = "sv", verbose)
}
\arguments{
\item{beta.sva}{Methylation levels matrix, one row per CpG site, one column per sample.}

\item{covariates}{Covariates data frame to include in regression model,
one row per sample, one column per covariate (Default: NULL).}

\item{random.seed}{Value with which to seed the pseudo random number generator for reproducible results}

\item{mod}{model matrix form \code{\link{svaPrep}}}

\item{n.sv}{Number of surrogate variables}

\item{mode}{Do sva or isva analysis, accepts values "sv" or "isv"}

\item{verbose}{Set verbosity of isva function}
}
\description{
Constructs function calls to surrogate variable analysis or independent surrogate variable analysis
}
