% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{test.pairwise.associations}
\alias{test.pairwise.associations}
\title{Tests associations between
between each column of y and each column of x.
Each column of y must be numeric.
In cases where a column of x contains factors,
the y-values for each factor level is compared
to the set of all y-values that are not outliers.
Outliers are identified as those below 1.5
times the inter-quartile range below the first quartile
or 1.5 times the inter-quartile range above the third quartile.}
\usage{
test.pairwise.associations(y, x)
}
\description{
Tests associations between
between each column of y and each column of x.
Each column of y must be numeric.
In cases where a column of x contains factors,
the y-values for each factor level is compared
to the set of all y-values that are not outliers.
Outliers are identified as those below 1.5
times the inter-quartile range below the first quartile
or 1.5 times the inter-quartile range above the third quartile.
}

