% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.r
\name{adjust.columns}
\alias{adjust.columns}
\title{Fit linear models with each column of \code{y}
as dependent variables and the fixed and random
effects as independent variables.
Return the residual matrix.}
\usage{
adjust.columns(y, fixed.effects = NULL, random.effects = NULL)
}
\description{
Fit linear models with each column of \code{y}
as dependent variables and the fixed and random
effects as independent variables.
Return the residual matrix.
}

