% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-frame.r
\name{check.data.frame}
\alias{check.data.frame}
\title{check that data frame \code{x} contains the required
columns and types or contents.
\code{columns} is a list whose names correspond to column names
and the values of each item is either a type (e.g. "character", "numeric")
or a vector of possible values.}
\usage{
check.data.frame(x, columns)
}
\description{
check that data frame \code{x} contains the required
columns and types or contents.
\code{columns} is a list whose names correspond to column names
and the values of each item is either a type (e.g. "character", "numeric")
or a vector of possible values.
}

