% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{guess.batch.vars}
\alias{guess.batch.vars}
\title{Guess which columns in sample sheet are batch variables}
\usage{
guess.batch.vars(norm.objects)
}
\arguments{
\item{norm.objects}{Output from \link{meffil.normalize.quantiles}}
}
\value{
Array of variable names
}
\description{
Guess which columns in sample sheet are batch variables
}
\examples{
\dontrun{

}
}
