% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifest.r
\name{guess.chip}
\alias{guess.chip}
\title{select a compatible chip the \code{object}
which may be an 'rg' object (see \code{\link{read.rg}()}
or a matrix (typically beta or methylation or unmethylation matrices)
with row names corresponding to feature/probe names.}
\usage{
guess.chip(object, chip = NA)
}
\description{
select a compatible chip the \code{object}
which may be an 'rg' object (see \code{\link{read.rg}()}
or a matrix (typically beta or methylation or unmethylation matrices)
with row names corresponding to feature/probe names.
}

