% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/knit-report.r
\name{knit.report}
\alias{knit.report}
\title{Knit report using working environment}
\usage{
knit.report(input.filename, output.filename, ...)
}
\arguments{
\item{output.filename}{Markdown or HTML output file.  An HTML file
is specified using the .htm, .html, .HTM or .HTML file extension.
When html is specified, a similarly named markdown file is also
generated.
All output files including cache and figures will appear in the
same folder as \code{output.filename}.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}

\item{rmd.filename}{RMD file.}
}
\description{
Knit report using working environment
}

