% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.add.featureset}
\alias{meffil.add.featureset}
\title{Add a feature set.}
\usage{
meffil.add.featureset(name, features)
}
\arguments{
\item{name}{Name of the new feature set.}

\item{features}{A data frame listing and describing all features.}
}
\value{
Assuming that \code{features} contains a satisfactory
set of columns, a new feature set is made available.
Thus, \code{name} will be added to the vector returned by
\code{\link{meffil.list.featuresets}()}.

The \code{features} data frame must contain the following columns:
\itemize{
\item{"name"}{character}
\item{"target"}{character}
\item{"type"}{values "i","ii" or "control"}
\item{"chromosome"}{values "chr1"-"chr22", "chrX" or "chrY"}
\item{"position"}{integer}
\item{"gene.symbol"}{character},
\item{"gene.accession"}{character},
\item{"gene.region"}{character},
\item{"cpg.island.name"}{character}
\item{"relation.to.island"}{character}
\item{"snp.exclude"}{logical}
}
}
\description{
Add a feature set.
}

