% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.controls}
\alias{meffil.cnv.controls}
\title{Create controls for calculating CNVs}
\usage{
meffil.cnv.controls(verbose = FALSE)
}
\arguments{
\item{verbose}{Default = FALSE>}
}
\value{
List with pre-calculated CNV data
}
\description{
Create controls for calculating CNVs
}

