% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.controls}
\alias{meffil.cnv.controls}
\title{Create controls for calculating CNVs}
\usage{
meffil.cnv.controls(rgset, verbose = FALSE)
}
\arguments{
\item{rgset}{An rgset of samples to be used as controls}

\item{verbose}{Default = FALSE}
}
\value{
List with intensities, sex, pre-calculated median values for use in CNV calling
}
\description{
Create controls for calculating CNVs
}

