% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-matrix.r
\name{meffil.cnv.matrix}
\alias{meffil.cnv.matrix}
\title{Create matrix of CNV values}
\usage{
meffil.cnv.matrix(cnv, featureset = "450k")
}
\arguments{
\item{cnv}{Output from \code{\link{meffil.calculate.cnv}()}.}

\item{featureset}{Name from \code{\link{meffil.list.featuresets}()} (Default: "450k").}
}
\value{
Matrix of ncpg x nsample
}
\description{
Create matrix of CNV values
}
