% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.matrix}
\alias{meffil.cnv.matrix}
\title{Create matrix of CNV values}
\usage{
meffil.cnv.matrix(cnv, featureset = "450k")
}
\arguments{
\item{cnv}{Output from \code{\link{meffil.calculate.cnv}()}.}

\item{featureset}{Name from \code{\link{meffil.list.featuresets}()}.}
}
\value{
Matrix of ncpg x nsample
}
\description{
Create matrix of CNV values
}

