% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cnv.r
\name{meffil.cnv.matrix}
\alias{meffil.cnv.matrix}
\title{Create matrix of CNV values}
\usage{
meffil.cnv.matrix(cnv)
}
\arguments{
\item{cnv}{Output from \code{meffil.calculate.cnv}}
}
\value{
Matrix of ncpg x nsample
}
\description{
Create matrix of CNV values
}

