% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functional-normalization.r
\name{meffil.compute.normalization.object}
\alias{meffil.compute.normalization.object}
\title{Normalization object}
\usage{
meffil.compute.normalization.object(basename, number.quantiles = 500,
  dye.intensity = 5000, probes = meffil.probe.info())
}
\arguments{
\item{basename}{IDAT file basename (see \code{\link{meffil.basenames}}).}

\item{number.quantiles}{Number of quantiles to compute for probe subset (Default: 500).}

\item{dye.intensity}{Reference intensity for scaling each color channel (Default: 5000).}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}
}
\value{
List containing control probe information, probe summaries
and quantiles.
}
\description{
Create a normalization object for a given Infinium HumanMethylation450 BeadChip.
}

