% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{meffil.control.matrix}
\alias{meffil.control.matrix}
\title{Infinium HumanMethylation450 BeadChip control matrix}
\usage{
meffil.control.matrix(qc.objects)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}
}
\value{
Matrix with one row per object consisting of control probe intensities and summaries.
}
\description{
Matrix containing control probe intensities from the Infinium HumanMethylation450 BeadChip.
}

