% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/normalize-objects.r
\name{meffil.control.matrix}
\alias{meffil.control.matrix}
\title{Infinium HumanMethylation450 BeadChip control matrix}
\usage{
meffil.control.matrix(objects)
}
\arguments{
\item{objects}{A list of outputs from \code{\link{meffil.compute.normalization.object}()}.}
}
\value{
Matrix with one column per object consisting of control probe intensities and summaries.
}
\description{
Matrix containing control probe intensities from the Infinium HumanMethylation450 BeadChip.
}

