% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{meffil.control.matrix}
\alias{meffil.control.matrix}
\title{Infinium HumanMethylation450 BeadChip control matrix}
\usage{
meffil.control.matrix(
  qc.objects,
  normalize = F,
  fixed.effects = NULL,
  random.effects = NULL
)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{normalize}{If \code{TRUE}, then control matrix is scaled and specified
fixed and random effects removed from the matrix.  Otherwise, the raw control matrix is returned.
(Default: \code{FALSE}).}

\item{fixed.effects}{Names of columns in samplesheet that should be included as fixed effects
along with control matrix principal components (Default: NULL).}

\item{random.effects}{Names of columns in samplesheet that should be included as random effects
(Default: NULL).}
}
\value{
Matrix with one row per object consisting of control probe intensities and summaries.
}
\description{
Matrix containing control probe intensities from the Infinium HumanMethylation450 BeadChip.
}
