% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/create-cell-type-reference.r
\name{meffil.create.cell.type.reference}
\alias{meffil.create.cell.type.reference}
\title{Create a cell type reference object}
\usage{
meffil.create.cell.type.reference(basenames, cell.types,
  number.quantiles = 500, number.sites = 50, probes = meffil.probe.info(),
  temp.dir = NULL, verbose = F)
}
\arguments{
\item{cell.types}{Vector of cell type names corresponding to sample \code{basename}s.}

\item{number.quantiles}{Length of numeric sequence to specify probe intensity distributions
(Default: 500).}

\item{number.sites}{Number of probes to characterise cell type methylation (Default: 50).
For each cell type, this number of probes with greater methylation than other cell types
and the same number with lesser methylation than the other cell types will be included.}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}

\item{basename}{IDAT file basename (see \code{\link{meffil.basenames}()}).}
}
\value{
A list specifying a cell type reference object that can be used by
\code{\link{meffil.estimate.cell.counts}()} can use to estimate cell counts.
The object is a list containing:
- \code{beta} The normalized methylation values of sites
differentially methylated between cell types.
- \code{norm.objects} The normalization objects used to compute \code{beta}
using \code{\link{meffil.normalize.objects}()} corresponding to the IDAT files
identified by \code{basenames}.
- \code{quantiles} The average quantiles of methylated and unmethylated signals
of probe sets defined by \code{subsets} (see below).
e.g. \code{quantiles[[name]]$M} provides the quantiles (\code{number.quantiles} quantiles)
of the probes specified by \code{subsets[[name]]}.
- \code{subsets} Probes on the microarray partitioned by relationship to CpG islands,
either in an island, in a shore or far from an island.
}
\description{
Create a cell type reference object for estimating cell counts
with the Infinium HumanMethylation450 BeadChip.
}

