% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/create-qc-object.r
\name{meffil.create.qc.object}
\alias{meffil.create.qc.object}
\title{Quality control object}
\usage{
meffil.create.qc.object(samplesheet.row, number.quantiles = 500,
  dye.intensity = 5000, verbose = F, detection.threshold = 0.01,
  bead.threshold = 3, sex.cutoff = -2, cell.type.reference = NULL)
}
\arguments{
\item{samplesheet.row}{Row from the data frame containing IDAT file and sample info (see \code{\link{meffil.read.samplesheet}} or \code{\link{meffil.create.samplesheet}}).}

\item{number.quantiles}{Number of quantiles to compute for probe subset (Default: 500).}

\item{dye.intensity}{Reference intensity for scaling each color channel (Default: 5000).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}

\item{detection.threshold}{Default value = 0.01.
All probes above this detection threshold detected.}

\item{bead.threshold}{Default value = 3.
All probes with less than this number of beads detected.}

\item{sex.cutoff}{Sex prediction cutoff. Default value = -2.}

\item{cell.type.reference}{Character string name of the cell type reference
to use for estimating cell counts. Estimates are not generated if set to NULL (default).
See \code{\link{meffil.get.cell.type.references}()} for a list of available
references.  New references can be created using
\code{\link{meffil.create.cell.type.reference}()}.}
}
\value{
List containing control probe information, probe summaries
and quantiles.  We call this a "QC object".
}
\description{
Create a quality control object for a given Infinium HumanMethylation450 BeadChip.
}

