% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-objects.r
\name{meffil.design.matrix}
\alias{meffil.design.matrix}
\title{Infinium HumanMethylation450 BeadChip normalization design matrix}
\usage{
meffil.design.matrix(objects, number.pcs)
}
\arguments{
\item{objects}{A list of outputs from \code{\link{meffil.compute.normalization.object}()}.}

\item{number.pcs}{Number of principal components to include in the design matrix (Default: \code{length(objects)}).}
}
\value{
Design matrix with one column for each of the first \code{number.pcs} prinicipal
components.
}
\description{
Design matrix derived by applying principal components analysis to control probes.
}

