% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-cell-counts.r
\name{meffil.estimate.cell.counts}
\alias{meffil.estimate.cell.counts}
\title{Estimate cell counts from a reference}
\usage{
meffil.estimate.cell.counts(qc.object, cell.type.reference, verbose = T)
}
\arguments{
\item{cell.type.reference}{Character string name of the cell type reference
to use for estimating cell counts.
See \code{\link{meffil.list.cell.type.references}()} for a list of available
references.  New references can be created using
\code{\link{meffil.add.cell.type.reference}()}.}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}

\item{object}{An object created by \code{\link{meffil.create.qc.object}()}.}
}
\value{
A list:
\itemize{
\item \code{counts} Cell count estimates.
\item \code{beta} Normalized methylation levels of sites used to differentiate
\item \code{reference} Name of the cell type reference used.
between reference cell types.
}

Results should be nearly identical to \code{\link[minfi]{estimateCellCounts}()}.
}
\description{
Estimate cell type ratios from methylation profiles of purified cell populations
(Infinium HumanMethylation450 BeadChip) using the the Houseman algorithm
(PMID 22568884).
}
