% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ewas.r
\name{meffil.ewas.qq.plot}
\alias{meffil.ewas.qq.plot}
\title{QQ plot for EWAS}
\usage{
meffil.ewas.qq.plot(ewas.object, sig.threshold = 1e-07, sig.color = "red",
  title = "QQ plot", xlab = bquote(-log[10]("expected p-values")),
  ylab = bquote(-log[10]("observed p-values")), lambda.method = "median")
}
\arguments{
\item{ewas.object}{Return object from \code{\link{meffil.ewas()}}.}

\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{sig.color}{Color for points corresponding to significant tests (Default: "red").}

\item{title}{Title for the plot (Default: "QQ plot").}

\item{xlab}{Label for the x-axis (Default: -log_10(expected p-values)).}

\item{ylab}{Label for the y-axis (Default: -log_10(observed p-values)).}

\item{lambda.method}{Method for calculating genomic inflation lambda.
Valid values are "median" or "regression" (Default: "median").}
}
\value{
List of \code{\link{ggplot}} for each analysis in \code{ewas.object}.
}
\description{
QQ plot for EWAS
}
