% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{meffil.ewas}
\alias{meffil.ewas}
\title{Epigenome-wide association study}
\usage{
meffil.ewas(beta, variable, covariates = NULL, batch = NULL,
  weights = NULL, cell.counts = NULL, isva = T, sva = T, n.sv = NULL,
  isva0 = F, isva1 = F, winsorize.pct = 0.05, outlier.iqr.factor = NA,
  most.variable = min(nrow(beta), 50000), featureset = NA,
  random.seed = 20161123, verbose = F)
}
\arguments{
\item{beta}{Methylation levels matrix, one row per CpG site, one column per sample.}

\item{variable}{Independent variable vector.}

\item{covariates}{Covariates data frame to include in regression model,
one row per sample, one column per covariate (Default: NULL).}

\item{batch}{Batch vector to be included as a random effect (Default: NULL).}

\item{weights}{Non-negative observation weights.
Can be a numeric matrix of individual weights of same dimension as \code{beta},
or a numeric vector of weights with length \code{ncol(beta)},
or a numeric vector of weights with length \code{nrow(beta)}.}

\item{cell.counts}{Proportion of cell counts for one cell type in cases
where the samples are mainly composed of two cell types (e.g. saliva) (Default: NULL).}

\item{isva}{Apply Independent Surrogate Variable Analysis (ISVA) to the
methylation levels and include the resulting variables as covariates in a
regression model (Default: TRUE).}

\item{sva}{Apply Surrogate Variable Analysis (SVA) to the
methylation levels and covariates and include
the resulting variables as covariates in a regression model (Default: TRUE).}

\item{n.sv}{Number of surrogate variables to calculate (Default: NULL).}

\item{winsorize.pct}{Apply all regression models to methylation levels
winsorized to the given level. Set to NA to avoid winsorizing (Default: 0.05).}

\item{outlier.iqr.factor}{For each CpG site, prior to fitting regression models,
set methylation levels less than
\code{Q1 - outlier.iqr.factor * IQR} or more than
\code{Q3 + outlier.iqr.factor * IQR} to NA. Here IQR is the inter-quartile
range of the methylation levels at the CpG site, i.e. Q3-Q1.
Set to NA to skip this step (Default: NA).}

\item{most.variable}{Apply (Independent) Surrogate Variable Analysis to the 
given most variable CpG sites (Default: 50000).}

\item{featureset}{Name from \code{\link{meffil.list.featuresets}()}  (Default: NA).}

\item{verbose}{Set to TRUE if status updates to be printed (Default: FALSE).}
}
\description{
Test association with each CpG site.
}
