% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.r
\name{meffil.ewas}
\alias{meffil.ewas}
\title{Epigenome-wide association study}
\usage{
meffil.ewas(beta, variable, covariates = NULL, batch = NULL,
  save.beta = 100, cell.counts = NULL, isva0 = T, isva1 = T,
  winsorize.pct = 0.05, most.variable = min(nrow(beta), 20000),
  verbose = F)
}
\arguments{
\item{beta}{Methylation levels matrix, one row per CpG site, one column per sample.}

\item{variable}{Independent variable vector.}

\item{covariates}{Covariates data frame to include in regression model,
one row per sample, one column per covariate (Default: NULL).}

\item{batch}{Batch vector to be included as a random effect (Default: NULL).}

\item{save.beta}{Number of most strongly associated CpG sites for which to retain
methylation levels for later plotting (Default: 100).}

\item{cell.counts}{Proportion of cell counts for one cell type in cases
where the samples are mainly composed of two cell types (e.g. saliva) (Default: NULL).}

\item{isva0}{Apply Independent Surrogate Variable Analysis (ISVA) to the
methylation levels and include the resulting variables as covariates in a
regression model (Default: TRUE).}

\item{isva1}{Apply Independent Surrogate Variable Analysis (ISVA) to the
methylation levels, covariates and \code{isva0} variables and include
the resulting variables as covariates in a regression model (Default: TRUE).}

\item{winsorize.pct}{Apply all regression models to methylation levels
winsorized to the given level (Default: 0.05).}

\item{most.variable}{Apply Independent Surrogate Variable Analysis to the 
given most variable CpG sites (Default: 20000).}
}
\description{
Test association with each CpG site.
}

