% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.featureset}
\alias{meffil.featureset}
\title{Obtain a list of features in a feature set.}
\usage{
meffil.featureset(featureset = "450k")
}
\arguments{
\item{featureset}{Name returned by \code{\link{meffil.list.featuresets}()} (Default: "450k").}
}
\value{
A data frame with one row for each feature.
}
\description{
Obtain a list of features in a feature set.
}
\examples{
x <- meffil.featureset("450k")

}
