% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-beta.r
\name{meffil.get.beta}
\alias{meffil.get.beta}
\title{Infinium HumanMethylation450 BeadChip methylation levels}
\usage{
meffil.get.beta(mu, pseudo = 100)
}
\arguments{
\item{mu}{Methylated/unmethylated signal from \code{\link{meffil.rg.to.mu}()}
or \code{\link{meffil.normalize.sample}()}.}

\item{pseudo}{Value to add to the denominator to make the methylation estimate more stable.}
}
\value{
Vector of 0..1 methylation level estimates.
Equal to methylated/(methylated + unmethylated + pseudo).
}
\description{
Compute beta values (methylation levels) from methylated/unmethylated signals
}

