% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-subsets.r
\name{meffil.get.y.sites}
\alias{meffil.get.y.sites}
\title{Get names of chromosome Y CpG sites in the feature set.}
\usage{
meffil.get.y.sites(featureset = "450k")
}
\description{
Get names of chromosome Y CpG sites in the feature set.
}

