% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifest.r
\name{meffil.list.featuresets}
\alias{meffil.list.featuresets}
\title{List of feature sets available.}
\usage{
meffil.list.featuresets()
}
\description{
By default, there is '450k', 'epic' and 'common'.
The 'common' feature set contains features in common to both the
'450k' and 'epic' feature sets.
This feature set can be used to handle datasets
with mixed EPIC and HumanMethylation450 microarrays.
}
\details{
In most cases, a feature corresponds to the two probes
from which it's value is derived.  Each CpG represented
on the chip for example
corresponds to a single feature derived from a probe measuring
methylated signal and a second probe measuring unmethylated signal.

Each control feature corresponds to a unique control probe.
}

