% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation-pcs.r
\name{meffil.methylation.pcs}
\alias{meffil.methylation.pcs}
\title{Compute principal components of a methylation matrix.}
\usage{
meffil.methylation.pcs(
  beta,
  probe.range = 50000,
  sites = NULL,
  samples = NULL,
  autosomal = T,
  winsorize.pct = NA,
  outlier.iqr.factor = NA,
  full.obj = F,
  verbose = F
)
}
\arguments{
\item{beta}{Output from \code{\link{meffil.normalize.samples}()},
either a matrix or a GDS filename.}

\item{probe.range}{Default = 50000. How many probes to be used in calculating PCs.}

\item{sites}{Subset of CpG sites to consider (row names of beta) (Default: NULL).}

\item{samples}{Subset of samples to consider (column names of beta) (Default: NULL).}

\item{autosomal}{If true, remove probes on sex chromosomes (Default: TRUE).}

\item{winsorize.pct}{Apply to methylation levels
winsorized to the given level. Set to NA to avoid winsorizing (Default: NA).}

\item{outlier.iqr.factor}{Apply to methylation after setting,
for each CpG site, values less than
\code{Q1 - outlier.iqr.factor * IQR} or more than
\code{Q3 + outlier.iqr.factor * IQR} to NA. Here IQR is the inter-quartile
range of the methylation levels at the CpG site, i.e. Q3-Q1.
Set to NA to skip this step (Default: NA).}

\item{full.obj}{Default = FALSE.  If true, then return the full `prcomp` object rather than just the PCs.}

\item{verbose=T}{Print progress messages?}
}
\value{
the principal components of \code{normalized.beta}.
}
\description{
Compute principal components of a methylation matrix.
}
