% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/most-variable.r
\name{meffil.most.variable.cpgs}
\alias{meffil.most.variable.cpgs}
\title{Most variable CpG sites}
\usage{
meffil.most.variable.cpgs(x, n = 1000, autosomal = T)
}
\arguments{
\item{x}{Methylation matrix (rows = CpG sites, columns = samples).}

\item{n}{Number of CpG sites to return.}

\item{autosomal}{If \code{TRUE}, then consider only autosomal CpG sites (Default: TRUE).}
}
\value{
The \code{n} CpG site identifiers (rownames of \code{x}) with the greatest variance in \code{x}.
}
\description{
Returns the most variable CpG sites (rows) in the methylation matrix.
}

