% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.normalization.parameters}
\alias{meffil.normalization.parameters}
\title{Specify parameters for testing normalization}
\usage{
meffil.normalization.parameters(
  norm.objects,
  variables = guess.batch.vars(norm.objects),
  control.pcs = 1:10,
  batch.pcs = 1:10,
  batch.threshold = 1e-50,
  colours = NULL
)
}
\arguments{
\item{norm.objects}{Output from \link{meffil.normalize.quantiles}}

\item{variables}{Default = guess.batch.vars(norm). Which variables in sample sheet to test}

\item{control.pcs}{Default = 1:10. Number of control PCs to test against batch variables}

\item{colours}{Colours to use for scatterplots.}
}
\value{
List of parameters
}
\description{
Specify parameters for testing normalization
}
\examples{
\dontrun{

}
}
