% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.normalization.summary}
\alias{meffil.normalization.summary}
\title{Perform tests to check normalization performance}
\usage{
meffil.normalization.summary(norm.objects, pcs,
  parameters = meffil.normalization.parameters(norm.objects),
  variables = NULL, verbose = TRUE)
}
\arguments{
\item{norm.objects}{Output from \link{meffil.normalize.quantiles}}

\item{pcs}{Output from \code{\link{meffil.methylation.pcs}()}
applied to the normalized methylation matrix
corresponding to \code{norm.objects}.}

\item{parameters}{Default = meffil.post.parameters(norm.objects). Report parameters.}

\item{variables}{Default = NULL. Data frame of variables to compare to
principal components (\code{pcs}).  Must contain \code{length(norm.objects)} rows.
Columns that are not factors are ignored.}

\item{verbose}{Default = TRUE}
}
\value{
List of tables and graphs.
}
\description{
Creates scree plot of PCs of control probes, tests for association of control probe PCs with batch variables, tests for association of normalized probes with batch variables, creates PCA plots
}
\examples{
\dontrun{

}
}

