% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.normalization.summary}
\alias{meffil.normalization.summary}
\title{Perform tests to check normalization performance}
\usage{
meffil.normalization.summary(normalized.beta, norm.objects,
  parameters = meffil.normalization.parameters(norm.objects),
  verbose = TRUE)
}
\arguments{
\item{normalized.beta}{Output from \code{meffil.normalize.samples}}

\item{parameters}{Default = meffil.post.parameters(norm) List of parameters.}

\item{verbose}{Default = TRUE}

\item{norm}{Output from \link{meffil.normalize.quantiles}}
}
\value{
List of tables and graphs
}
\description{
Creates scree plot of PCs of control probes, tests for association of control probe PCs with batch variables, tests for association of normalized probes with batch variables
}
\examples{
\dontrun{

}
}

