% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-dataset.r
\name{meffil.normalize.dataset}
\alias{meffil.normalize.dataset}
\title{Functional normalization}
\usage{
meffil.normalize.dataset(samplesheet, number.quantiles = 500,
  detection.threshold = 0.01, bead.threshold = 3, sex.cutoff = -2,
  qc.parameters = meffil.qc.parameters(), qc.file = "meffil-qc-report.md",
  author = "Analyst", study = "IlluminaHuman450 data", number.pcs = 2,
  beta = T, pseudo = 100, temp.dir = NULL, norm.parameters = NULL,
  norm.file = "meffil-normalization-report.md", verbose = FALSE)
}
\arguments{
\item{samplesheet}{Output from \code{\link{meffil.read.samplesheet}()} or
\code{\link{meffil.create.samplesheet}()}.

Arguments to \code{\link{meffil.qc}()}:}

\item{number.quantiles}{}

\item{detection.threshold}{}

\item{bead.threshold}{}

\item{sex.cutoff}{Argument to \code{\link{meffil.qc.summary}()}:}

\item{qc.parameters}{(parameters)

Arguments to \code{\link{meffil.qc.report}()}:}

\item{qc.file}{(output.file)}

\item{author}{}

\item{study}{Arguments to \code{\link{meffil.normalize.quantiles}()}:}

\item{number.pcs}{Arguments to \code{\link{meffil.normalize.samples}()}:}

\item{beta}{}

\item{pseudo}{}

\item{temp.dir}{Arguments to \code{\link{meffil.normalization.summary}()}:}

\item{norm.parameters}{(parameters)}

\item{norm.file}{(output.file)

Other:}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}
}
\value{
A list:
- qc.summary \code{\link{meffil.qc.summary}()} output.
- norm \code{\link{meffil.normalize.quantiles}()} output.
- data \code{\link{meffil.normalize.samples}()} output.
- norm.summary \code{\link{meffil.normalization.summary}()} output.
}
\description{
Apply functional normalization to a set of Infinium HumanMethylation450 BeadChip IDAT files.
}
\details{
Fortin JP, Labbe A, Lemire M, Zanke BW, Hudson TJ, Fertig EJ, Greenwood CM, Hansen KD.
Functional normalization of 450k methylation array data
improves replication in large cancer studies.
Genome Biol. 2014 Dec 3;15(12):503. doi: 10.1186/s13059-014-0503-2.
PMID: 25599564
}

