% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-dataset.r
\name{meffil.normalize.dataset}
\alias{meffil.normalize.dataset}
\title{Functional normalization}
\usage{
meffil.normalize.dataset(samplesheet, number.quantiles = 500,
  detection.threshold = 0.01, bead.threshold = 3, sex.cutoff = -2,
  chip = NA, featureset = chip, cell.type.reference = NA,
  qc.parameters = meffil.qc.parameters(), qc.file = "meffil-qc-report.md",
  author = "Analyst", study = "IlluminaHuman450 data", number.pcs = 2,
  fixed.effects = NULL, random.effects = NULL, pseudo = 100,
  just.beta = T, probe.range = 5000, autosomal = T,
  norm.parameters = NULL, norm.file = "meffil-normalization-report.md",
  verbose = FALSE)
}
\arguments{
\item{samplesheet}{Output from \code{\link{meffil.read.samplesheet}()} or
\code{\link{meffil.create.samplesheet}()}.

Arguments to \code{\link{meffil.qc}()}:}

\item{cell.type.reference}{Argument to \code{\link{meffil.qc.summary}()}:}

\item{qc.parameters}{(parameters)

Arguments to \code{\link{meffil.qc.report}()}:}

\item{qc.file}{(output.file)}

\item{study}{Arguments to \code{\link{meffil.normalize.quantiles}()}:}

\item{fixed.effects}{Names of columns in samplesheet that should be included as fixed effects
along with control matrix principal components (Default: NULL).}

\item{random.effects}{Names of columns in samplesheet that should be included as random effects
(Default: NULL).

Arguments to \code{\link{meffil.normalize.samples}()}:}

\item{pseudo}{Arguments to \code{\link{meffil.methylation.pcs}()}.}

\item{probe.range}{(Default: 5000).}

\item{autosomal}{(Default: TRUE).

Arguments to \code{\link{meffil.normalization.summary}()}:}

\item{norm.parameters}{(parameters)}

\item{norm.file}{(output.file)

Other:}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution
(Default: \code{FALSE}).}

\item{npcs}{(Default: 1:10).}
}
\value{
A list:
- qc.summary \code{\link{meffil.qc.summary}()} output.
- norm \code{\link{meffil.normalize.quantiles}()} output.
- beta Normalized beta matrix (methylation levels).
- norm.summary \code{\link{meffil.normalization.summary}()} output.
}
\description{
Apply functional normalization to a set of Infinium HumanMethylation450 BeadChip IDAT files.
}
\details{
Fortin JP, Labbe A, Lemire M, Zanke BW, Hudson TJ, Fertig EJ, Greenwood CM, Hansen KD.
Functional normalization of 450k methylation array data
improves replication in large cancer studies.
Genome Biol. 2014 Dec 3;15(12):503. doi: 10.1186/s13059-014-0503-2.
PMID: 25599564
}
