% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-quantiles.r
\name{meffil.normalize.quantiles}
\alias{meffil.normalize.quantiles}
\title{Normalize microarray quantiles}
\usage{
meffil.normalize.quantiles(qc.objects, number.pcs = 2, verbose = F)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{number.pcs}{Number of control matrix principal components to adjust for (Default: 2).}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
Same list as input with additional elements added for each sample
including normalized quantiles needed for normalizing each sample.
}
\description{
Normalize microarray quantiles using controls extracted (Infinium HumanMethylation450 BeadChip).
}

