% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-sample.r
\name{meffil.normalize.sample}
\alias{meffil.normalize.sample}
\title{Normalize Infinium HumanMethylation450 BeadChip}
\usage{
meffil.normalize.sample(norm.object, verbose = F)
}
\arguments{
\item{norm.object}{An element of \code{\link{meffil.normalize.quantiles}()}.}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
List containing normalized methylated and unmethylated signals.
}
\description{
Normalize sample methylation data using normalized quantiles.
}
