% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-samples.r
\name{meffil.normalize.samples}
\alias{meffil.normalize.samples}
\title{Normalize Infinium HumanMethylation450 BeadChips}
\usage{
meffil.normalize.samples(norm.objects, pseudo = 100, just.beta = T,
  cpglist.remove = NULL, filename = NULL, max.bytes = 2^30 - 1,
  verbose = F, ...)
}
\arguments{
\item{norm.objects}{The list or sublist of \code{\link{meffil.normalize.quantiles}()}.}

\item{pseudo}{Value to add to the denominator to make the methylation
estimate more stable when calculating methylation levels (Default: 100).}

\item{just.beta}{If \code{TRUE}, then return just the normalized methylation levels; otherwise,
return the normalized methylated and unmethylated matrices (Default: TRUE).}

\item{cpglist.remove}{Optional list of CpGs to exclude from final output}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}.}
}
\value{
If \code{just.beta == TRUE}, the normalized matrix of 
methylation levels between between 0 and 1
equal to methylated signal/(methylated + unmethylated signal + pseudo).
Otherwise, a list containing two matrices, the normalized methylated and unmethylated signals.
}
\description{
Normalize a set of samples using their normalized quality control objects.
}

