% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-samples.r
\name{meffil.normalize.samples}
\alias{meffil.normalize.samples}
\title{Normalize Infinium HumanMethylation450 BeadChips}
\usage{
meffil.normalize.samples(objects, beta = T, pseudo = 100,
  probes = meffil.probe.info(), verbose = F, temp.dir = tempdir(),
  cpglist.remove = NULL, ...)
}
\arguments{
\item{objects}{A list or sublist returned by \code{\link{meffil.normalize.objects}()}.}

\item{beta}{If \code{TRUE} (default), then the function returns
the normalized matrix of methylation levels; otherwise, it returns
the normalized matrices of methylated and unmethylated signals.}

\item{pseudo}{Value to add to the denominator to make the methylation
estimate more stable when calculating methylation levels (Default: 100).}

\item{probes}{Probe annotation used to construct the control matrix
(Default: \code{\link{meffil.probe.info}()}).}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{temp.dir}{Option to choose alternative temporary directory, otherwise uses result from \code{tempdir()}}

\item{cpglist.remove}{Optional list of CpGs to exclude from final output}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}
except for \code{ret.bytes}.}
}
\value{
Matrix of normalized methylation levels if \code{beta} is \code{TRUE};
otherwise matrices of normalized methylated and unmethylated signals.
Matrices returned have one column per sample and one row per CpG site.
Methylation levels are values between 0 and 1
equal to methylated signal/(methylated + unmethylated signal + pseudo).
}
\description{
Normalize a set of samples using their normalization objects.
}

