% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-samples.r
\name{meffil.normalize.samples}
\alias{meffil.normalize.samples}
\title{Normalize Infinium HumanMethylation450 BeadChips}
\usage{
meffil.normalize.samples(
  norm.objects,
  pseudo = 100,
  just.beta = T,
  cpglist.remove = NULL,
  remove.poor.signal = F,
  dup.fun = function(x) median(x, na.rm = T),
  max.bytes = 2^30 - 1,
  gds.filename = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{norm.objects}{The list or sublist of \code{\link{meffil.normalize.quantiles}()}.}

\item{pseudo}{Value to add to the denominator to make the methylation
estimate more stable when calculating methylation levels (Default: 100).}

\item{just.beta}{If \code{TRUE}, then return just the normalized methylation levels; otherwise,
return the normalized methylated and unmethylated matrices (Default: TRUE).}

\item{cpglist.remove}{Optional list of CpGs to exclude from final output}

\item{remove.poor.signal}{Set methylation values for poorly detected probes
to missing  (Default: \code{FALSE}). Poor signal was
identified during QC by \code{\link{meffil.qc}()}
as signal that failed to pass the
detection p-value threshold (\code{detection.threshold})
or bead threshold (\code{bead.threshold}).}

\item{dup.fun}{Function to collapse duplicate probes
(EPIC v2 has over 5000 duplicated probes). If NULL, then
duplicates are not collapsed (Default: median).}

\item{gds.filename}{If not \code{NULL} (default), then saves the output to a GDS (Genomic Data Structure).
This is for cases where the output is too large to fit into main memory.
The GDS option assumes that argument \code{just.beta == TRUE}.}

\item{verbose}{If \code{TRUE}, then detailed status messages are printed during execution (Default: \code{FALSE}).}

\item{...}{Arguments passed to \code{\link[parallel]{mclapply}()}.}
}
\value{
If \code{just.beta == TRUE}, the normalized matrix of
methylation levels between between 0 and 1
equal to methylated signal/(methylated + unmethylated signal + pseudo).
Otherwise, a list containing two matrices, the normalized methylated and unmethylated signals.
If \code{gds.filename} is not \code{NULL}, then the output is saved to the GDS file
rather than retained in memory and returned to the caller.
The library 'gdsfmt' must be installed in this case.
}
\description{
Normalize a set of samples using their normalized quality control objects.
}
