% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalize-objects.r
\name{meffil.pcs}
\alias{meffil.pcs}
\title{Calculate control probe PCs}
\usage{
meffil.pcs(objects)
}
\arguments{
\item{objects}{A list of outputs from \code{\link{meffil.compute.normalization.object}()}.}
}
\value{
PCA of control probes
}
\description{
Calculate control probe PCs
}

