% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/design-matrix.r
\name{meffil.pcs}
\alias{meffil.pcs}
\title{Calculate control probe PCs}
\usage{
meffil.pcs(qc.objects, fixed.effects = NULL, random.effects = NULL)
}
\arguments{
\item{qc.objects}{A list of outputs from \code{\link{meffil.create.qc.object}()}.}

\item{fixed.effects}{Names of columns in samplesheet that should be included as fixed effects
along with control matrix principal components (Default: NULL).}

\item{random.effects}{Names of columns in samplesheet that should be included as random effects
(Default: NULL).}
}
\value{
PCA of control probes
}
\description{
Calculate control probe PCs
}

