% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.beadnum.cpgs}
\alias{meffil.plot.beadnum.cpgs}
\title{Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample}
\usage{
meffil.plot.beadnum.cpgs(qc.objects, threshold = 0.05)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}

\item{threshold}{Cut off value for proportion of samples with poor detection p values. Default 0.05.}
}
\value{
Data frame of results plus plot
}
\description{
Manhattan plot of number of beads by probe - percentage of probes with beads < 3 for each sample
}
\examples{
\dontrun{

}
}

