% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.plot.beadnum.samples}
\alias{meffil.plot.beadnum.samples}
\title{Plot number of beads per sample}
\usage{
meffil.plot.beadnum.samples(qc.objects, threshold = 0.05,
  colour.code = NULL)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}

\item{threshold}{Cut off value for proportion of CpGs with low bead numbers. Default 0.05}

\item{colour.code}{Array of length n samples to colour code points. Defaults to NULL}
}
\value{
Data frame of results plus plot
}
\description{
Plot number of beads per sample
}
\examples{
\dontrun{

}
}

