% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.detectionp.samples}
\alias{meffil.plot.detectionp.samples}
\title{Plot detection p values from idat files}
\usage{
meffil.plot.detectionp.samples(qc.objects, threshold = 0.05,
  colour.code = NULL)
}
\arguments{
\item{qc.objects}{From \code{meffil.normalize.objects}}

\item{threshold}{Cut off value for proportion of CpGs with poor detection p values. Default 0.05}

\item{colour.code}{Array of length n samples to colour code points. Defaults to NULL}
}
\value{
Data frame of results plus plot
}
\description{
Plot detection p values from idat files
}
\examples{
\dontrun{

}
}

