% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-report.r
\name{meffil.plot.probe.batch}
\alias{meffil.plot.probe.batch}
\title{Test normalized betas for association with known batch variables}
\usage{
meffil.plot.probe.batch(
  norm.objects,
  pcs,
  variables = guess.batch.vars(norm.objects),
  additional = NULL,
  batch.threshold = 1e-50,
  cols = NULL,
  verbose = T
)
}
\arguments{
\item{norm.objects}{Output from \code{\link{meffil.normalize.quantiles}()}.}

\item{pcs}{Output from \code{\link{meffil.methylation.pcs}()}
applied to the normalized methylation matrix
corresponding to \code{norm.objects}.}

\item{variables}{Default = guess.batch.vars(norm). Which variables in sample sheet to test}

\item{additional.}{Default = NULL. Data frame containing variables to test for association
with control matrix PCs. Must have \code{nrow(additional) == length(norm.objects)}.}

\item{verbose=T}{Print progress messages?}
}
\value{
List of table of results and graph
}
\description{
Performs association of each of \code{n} PCs calculated from most variable CpG sites (after normalization) against each of \code{m} measured batch variables
}
\examples{
\dontrun{

}
}
