% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.plot.sex}
\alias{meffil.plot.sex}
\title{Plot predicted sex}
\usage{
meffil.plot.sex(samplesheet, norm.objects, outlier.sd = 3)
}
\arguments{
\item{samplesheet}{From \code{read.450k.sheet}}

\item{norm.objects}{From \code{meffil.normalize.objects}}
}
\value{
Data frame of results plus plot
}
\description{
Plot predicted sex
}
\examples{
\dontrun{

}
}

