% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe-info.r
\name{meffil.probe.info}
\alias{meffil.probe.info}
\title{Probe type and location annotation}
\usage{
meffil.probe.info()
}
\value{
The data frame being used to annotate
the Infinium HumanMethylation450 BeadChip.
The default uses
\code{\link[IlluminaHumanMethylation450k]{IlluminaHumanMethylation450kmanifest}}
extensively.  Probe locations are obtained using
\code{\link[IlluminaHumanMethylation450kanno.ilmn12.hg19]{IlluminaHumanMethylation450kanno.ilmn12.hg19}}.
}
\description{
Probe type and location annotation
}

