% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.qc.parameters}
\alias{meffil.qc.parameters}
\title{Specify parameters for QC}
\usage{
meffil.qc.parameters(colour.code = NULL, control.categories = NULL,
  sex.outlier.sd = 3, meth.unmeth.outlier.sd = 3,
  control.means.outlier.sd = 5, detectionp.samples.threshold = 0.05,
  beadnum.samples.threshold = 0.05, detectionp.cpgs.threshold = 0.05,
  beadnum.cpgs.threshold = 0.05)
}
\arguments{
\item{colour.code}{Default value = NULL <what param does>}

\item{control.categories}{Default value = control.probe.categories() <what param does>}

\item{sex.outlier.sd}{Default value = 3 <what param does>}

\item{meth.unmeth.outlier.sd}{Default value = 3 <what param does>}

\item{control.means.outlier.sd}{Default value = 5 <what param does>}

\item{detectionp.samples.threshold}{Default value = 0.05 <what param does>}

\item{beadnum.samples.threshold}{Default value = 0.05 <what param does>}

\item{detectionp.cpgs.threshold}{Default value = 0.05 <what param does>}

\item{beadnum.cpgs.threshold}{Default value = 0.05 <what param does>}
}
\value{
List of parameter values
}
\description{
Specify parameters for QC
}
\examples{
\dontrun{

}
}

