% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.qc.parameters}
\alias{meffil.qc.parameters}
\title{Specify parameters for QC}
\usage{
meffil.qc.parameters(colour.code = NULL, control.categories = NULL,
  sex.outlier.sd = 3, meth.unmeth.outlier.sd = 3,
  control.means.outlier.sd = 5, detectionp.samples.threshold = 0.2,
  beadnum.samples.threshold = 0.2, detectionp.cpgs.threshold = 0.2,
  beadnum.cpgs.threshold = 0.2, snp.concordance.threshold = 0.9,
  sample.genotype.concordance.threshold = 0.9)
}
\arguments{
\item{colour.code}{Default value = NULL <what param does>}

\item{control.categories}{Default value = control.probe.categories() <what param does>}

\item{sex.outlier.sd}{Sets the standard deviation multiple at which sex outliers are identified. Default value = 3.}

\item{meth.unmeth.outlier.sd}{Sets the standard deviation multiple at which methylated/unmethylated signal outliers are identified. Default value = 3.}

\item{control.means.outlier.sd}{Sets the standard deviation multiple at which control probe signals are identified as outliers. Default value = 5 <what param does>}

\item{detectionp.samples.threshold}{Detection p-value threshold.  Probes with values above this are considered undetected. Default value = 0.05}

\item{beadnum.samples.threshold}{A sample is excluded if the given proportion of probes has low bead number. Default value = 0.05}

\item{detectionp.cpgs.threshold}{A sample is excluded if the given proporition of probes are undetected. Default value = 0.05}

\item{beadnum.cpgs.threshold}{A probe is excluded if the given proportion of samples have low bead number. Default value = 0.05}

\item{snp.concordance.threshold}{Minimum required concordance between supplied genotypes and genotypes estimated from a SNP probe. Default value = 0.99}

\item{sample.genotype.concordance.threshold}{Minimum required concordance between supplied genotypes and genotypes estimated from SNP probes for a given individual. Default value = 0.9 <what param does>}
}
\value{
List of parameter values
}
\description{
Specify parameters for QC
}
\examples{
\dontrun{

}
}
