% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.r
\name{meffil.qc}
\alias{meffil.qc}
\title{Perform QC on HumanMethylation450 idat files}
\usage{
meffil.qc(samplesheet, number.quantiles = 500, dye.intensity = 5000,
  detection.threshold = 0.01, bead.threshold = 3, sex.cutoff = -2,
  verbose = F)
}
\arguments{
\item{samplesheet}{Data frame containing IDAT file and sample info (see \code{\link{meffil.read.samplesheet}} pr \code{\link{meffil.create.samplesheet}}).}

\item{number.quantiles}{Number of quantiles to compute for probe subset (Default: 500).}

\item{dye.intensity}{Reference intensity for scaling each color channel (Default: 5000).}

\item{detection.threshold}{Default value = 0.01.
All probes above this detection threshold detected.}

\item{bead.threshold}{Default value = 3.
All probes with less than this number of beads detected.}

\item{sex.cutoff}{Sex prediction cutoff. Default value = -2.}

\item{verbose}{If \code{TRUE}, then status messages are printed during execution (Default: \code{FALSE}).}
}
\value{
List containing control probe information, probe summaries
and quantiles.
}
\description{
Read in control matrices for each sample.
Perform background correction and R/G dye bias correction.
Predict sex
}

