% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pre-process-plots.r
\name{meffil.qc.report}
\alias{meffil.qc.report}
\title{Generate QC report}
\usage{
meffil.qc.report(qc.summary, output.file = "meffil.qc.report.html",
  author = "Analyst", studyname = "IlluminaHuman450 data", ...)
}
\arguments{
\item{qc.summary}{Output from \code{meffil.qc.summary}.}

\item{output.file}{Default = "meffil.qc.report.html"
If specified then a html report will be generated summarising the QC.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{studyname}{Default = "IlluminaHuman450 data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{rmarkdown::render}}}
}
\description{
Generate HTML file that summarises the QC.
}
\examples{
\dontrun{

}
}

