% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.qc.report}
\alias{meffil.qc.report}
\title{Generate QC report}
\usage{
meffil.qc.report(qc.summary, output.file = "meffil-qc-report.html",
  author = "Analyst", study = "IlluminaHuman450 data", ...)
}
\arguments{
\item{qc.summary}{Output from \code{meffil.qc.summary}.}

\item{output.file}{Default = "meffil.qc.report.html"
If specified then a html report will be generated summarising the QC.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{study}{Default = "IlluminaHuman450 data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\description{
Generate HTML file that summarises the QC.
}
\examples{
\dontrun{

}
}

