% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.qc.report}
\alias{meffil.qc.report}
\title{Generate QC report}
\usage{
meffil.qc.report(qc.summary, output.file = "qc-report.html",
  author = "Analyst", study = "Illumina methylation data", ...)
}
\arguments{
\item{qc.summary}{Output from \code{meffil.qc.summary}.}

\item{output.file}{Default = "meffil-qc-report.html".
If the file extension is not .htm, .html, .HTM or .HTML then
output will be in markdown format.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{study}{Default = "Illumina methylation data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\description{
Generate HTML file that summarises the QC.
}
\examples{
\dontrun{

}
}
