% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc-report.r
\name{meffil.qc.summary}
\alias{meffil.qc.summary}
\title{Perform QC analysis on idat files}
\usage{
meffil.qc.summary(qc.objects, parameters = meffil.qc.parameters(),
  verbose = TRUE)
}
\arguments{
\item{qc.objects}{From \code{meffil.qc}}

\item{parameters}{Default = meffil.qc.parameters(). List of parameter values. See \code{\link{meffil.qc.parameters}}}
}
\value{
List
}
\description{
Performs a number of QC analyses including checking for sex differences, methylated vs unmethylated levels,
deviation from control probe means, detection p-values and bead numbers per sample and probe.
}
\details{
Also returns list of sample IDs and CPGs that are low quality.
}
\examples{
\dontrun{

}
}

