% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/snp-betas.r
\name{meffil.snp.betas}
\alias{meffil.snp.betas}
\title{Matrix of SNP 'beta' values}
\usage{
meffil.snp.betas(qc.objects)
}
\arguments{
\item{qc.objects}{List of objects obtained from \code{\link{meffil.qc}()}
or \code{\link{meffil.create.qc.object}()}.}
}
\description{
Matrix of SNP 'beta' values
}

