% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-concordance.r
\name{meffil.snp.concordance}
\alias{meffil.snp.concordance}
\title{Concordance between genotypes and SNP betas}
\usage{
meffil.snp.concordance(snp.betas, genotypes, snp.threshold = 0.99,
  sample.threshold = 0.9)
}
\value{
Returns a list of two vectors:
one providing concordances between genotypes and SNP betas for matched samples,
a second providing concordances between genotypes and SNP betas for matched SNPs.
}
\description{
genotypes <- meffil.extract.genotypes(raw.filenames)
snp.betas <- meffil.snp.betas(qc.objects)
meffil.snp.concordance(snp.betas, genotypes[rownames(snp.betas),colnames(snp.betas)])
}

