% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rmd2html.r
\name{rmd2html}
\alias{rmd2html}
\title{Warning: It is quite likely that this will be called within an RMD file
implying a recursive call to knit(). this will generate "duplicate label"
errors for unlabelled chunks. To avoid this, all RMD files in this
package should contain only named chunks.
Supposedly this error can also be avoided by setting the following option:
     options(knitr.duplicate.label='allow')
I tried this but it didn't seem to help.}
\usage{
rmd2html(input.filename, output.filename, ...)
}
\description{
Warning: It is quite likely that this will be called within an RMD file
implying a recursive call to knit(). this will generate "duplicate label"
errors for unlabelled chunks. To avoid this, all RMD files in this
package should contain only named chunks.
Supposedly this error can also be avoided by setting the following option:
     options(knitr.duplicate.label='allow')
I tried this but it didn't seem to help.
}

